% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_areas}
\alias{assign_areas}
\title{Assign areas}
\usage{
assign_areas(major_stat_area_description, area_regex = c("3[CD]+",
  "5[AB]+", "5[CDE]+"))
}
\arguments{
\item{major_stat_area_description}{A vector of major statistical area
descriptions.}

\item{area_regex}{A vector of regular expressions describing the areas group.}
}
\description{
Assign areas
}
\examples{
x <- c("5D: NORTHERN HECATE STRAIT", "3C: S.W. VANCOUVER ISLAND", "3D: N.W. VANCOUVER ISLAND")
assign_areas(x)
}
