% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-samples.R
\name{bind_samples}
\alias{bind_samples}
\title{Join survey and commercial samples}
\usage{
bind_samples(dat_comm, dat_survey)
}
\arguments{
\item{dat_comm}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{dat_survey}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}
}
\description{
Join survey and commercial samples
}
\examples{
\dontrun{
dat_comm <- gfdata::get_commercial_samples("lingcod")
dat_survey <- gfdata::get_survey_samples("lingcod")
dat <- bind_samples(dat_comm, dat_survey)
}
}
