% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{cache_pbs_data}
\alias{cache_pbs_data}
\title{Cache all PBS groundfish data for one or more species}
\usage{
cache_pbs_data(species, path = "data-cache")
}
\arguments{
\item{species}{A character vector of species common names}

\item{path}{The folder where the cached data will be saved}
}
\description{
Cache all PBS groundfish data for one or more species
}
\seealso{
Other get PBS data functions: \code{\link{get_pbs_ageing_precision}},
  \code{\link{get_pbs_bioindex}},
  \code{\link{get_pbs_catch}},
  \code{\link{get_pbs_commsamples}},
  \code{\link{get_pbs_cpue_index}},
  \code{\link{get_pbs_cpue}},
  \code{\link{get_pbs_sample_trips}},
  \code{\link{get_pbs_strata}},
  \code{\link{get_pbs_survey}},
  \code{\link{get_pbs_survsamples}},
  \code{\link{get_sara_dat}}
}
