% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{calc_iphc_full_res}
\alias{calc_iphc_full_res}
\title{Do all the calculations for the IPHC survey.}
\usage{
calc_iphc_full_res(set_counts, sp)
}
\arguments{
\item{set_counts}{species-specific set-level counts from \code{\link[=tidy_iphc_survey]{tidy_iphc_survey()}}
or **other.}

\item{sp}{Species names (as used in gfplot).}
}
\value{
List containing

ser_longest: tibble for the longest time series that can be made for this
species, as output from \code{\link[=calc_iphc_ser_AB]{calc_iphc_ser_AB()}}; either Series A or AB.

full_coast: whether or not the longest time series can be considered
representative of the full coast (based on the paired t-tests).

ser_all: Series A, B, C and D, as output from \code{\link[=calc_iphc_ser_all]{calc_iphc_ser_all()}}.

test_AB: t-test results from \code{\link[=calc_iphc_ser_AB]{calc_iphc_ser_AB()}}

test_AD: t-test results from \code{\link[=compare_iphc_ser_A_D]{compare_iphc_ser_A_D()}}

test_BC: t-test results from \code{\link[=compare_iphc_ser_B_C]{compare_iphc_ser_B_C()}}

If no observations at all for the species then return NA.
}
\description{
From species-specific set-level counts, derive all the required Series and
test their equivalence.
}
