% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{fit_cpue_index}
\alias{fit_cpue_index}
\alias{predict_cpue_index}
\alias{plot_cpue_index}
\alias{tidy_cpue_index_coefs}
\alias{plot_cpue_index_coefs}
\alias{plot_cpue_index_jk}
\title{Commercial CPUE index standardization}
\usage{
fit_cpue_index(dat, formula_binomial = pos_catch ~ year_factor + f(month) +
  f(vessel) + f(locality) + f(depth) + f(latitude),
  formula_lognormal = log(spp_catch/hours_fished) ~ year_factor + f(month) +
  f(vessel) + f(locality) + f(depth) + f(latitude))

predict_cpue_index(object, center = FALSE)

plot_cpue_index(predicted_dat, all_models = TRUE)

tidy_cpue_index_coefs(object)

plot_cpue_index_coefs(object)

plot_cpue_index_jk(object, terms = c("f(month)", "f(vessel)", "f(locality)",
  "f(depth)", "f(latitude)"), return_data = FALSE)
}
\arguments{
\item{dat}{A data frame from \code{\link[=tidy_cpue_index]{tidy_cpue_index()}}, or a similarly
formatted data frame.}

\item{formula_binomial}{Formula for the binomial model.}

\item{formula_lognormal}{Formula for the lognormal model.}

\item{object}{Model output from \code{\link[=fit_cpue_index]{fit_cpue_index()}}.}

\item{center}{Should the index be centered by subtracting the mean in link space?}

\item{predicted_dat}{Input data frame, for example from \code{\link[=predict_cpue_index]{predict_cpue_index()}}.}

\item{all_models}{TODO}

\item{terms}{A character vector of terms to jackknife out.}

\item{return_data}{Logical: should the data be returned?}
}
\value{
A ggplot object

A ggplot object
}
\description{
\itemize{
\item \code{fit_cpue_index()} fits a delta-lognormal commercial CPUE
standardization model
}

\itemize{
\item \code{predict_cpue_index()} predicts from a delta-lognormal
commercial CPUE standardization model
}

\itemize{
\item \code{plot_cpue_index()} plots a delta-lognormal commercial CPUE
standardization model
}

\itemize{
\item \code{tidy_cpue_index_coefs()} extracts coefficients from a CPUE
index standardization model
}

\itemize{
\item \code{plot_cpue_index_coefs()} plots coefficients from a CPUE index
standardization model
}

\itemize{
\item \code{plot_cpue_index_jk()} "jackknifes" out terms one by one to
test the sensitivity of the standardization model to any one term.
}
}
\details{
Note that for coefficients for predictors treated as factors (i.e.
likely all of the predictors), the coefficients represent the difference
from the base level factor, which would be the first factor level
alpha-numerically. For example, month 02 represents the estimated
difference between February and January.
}
\examples{
# A simulated example:
set.seed(1)
d <- sim_cpue()
d

m <- fit_cpue_index(d,
  formula_binomial = pos_catch ~ year_factor + f(vessel),
  formula_lognormal = log(spp_catch / hours_fished) ~ year_factor + f(vessel)
)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

plot_cpue_index_jk(m, terms = "f(vessel)")

\dontrun{
# An example with PBS data:
# (the data extraction will be slow)

d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")

m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

plot_cpue_index_jk(m)
}
}
