% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{f}
\alias{f}
\alias{get_most_common_level}
\title{Force factors to be sequential within the positive or binary data sets for TMB}
\usage{
f(x, ref = get_most_common_level)

get_most_common_level(x)
}
\arguments{
\item{x}{A vector of factor/character values.}

\item{base}{A function to define the base "reference" level or a character
object definining the base level. Defaults to the most common level.}
}
\description{
This is necessary because sometimes, for example, a certain depth factor
level appears in the binary data set (because this depth was fished at), but
not in the positive data set (because a given species was now are caught at
that depth). This function also sets the base/reference level of the factor.
}
\examples{
a <- factor(c("a", "b", "c", "c"))
f(a)
f(a, "b")
}
