% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index-tweedie.R
\name{fit_cpue_index_glmmtmb}
\alias{fit_cpue_index_glmmtmb}
\title{Fit CPUE with glmmTMB}
\usage{
fit_cpue_index_glmmtmb(dat, formula = cpue ~ year_factor, ...)
}
\arguments{
\item{dat}{Data}

\item{formula}{Formula}

\item{...}{Other arguments to \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}}
}
\description{
Fit CPUE with glmmTMB
}
