% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index-tweedie.R
\name{fit_cpue_index_tweedie}
\alias{fit_cpue_index_tweedie}
\title{Commercial CPUE index standardization (Tweedie version)}
\usage{
fit_cpue_index_tweedie(dat, formula = cpue ~ year_factor)
}
\arguments{
\item{dat}{Data TODO}

\item{formula}{Formula TODO}
}
\description{
Commercial CPUE index standardization (Tweedie version)
}
\examples{
\donttest{
set.seed(1)
d <- sim_cpue()
d$cpue <- d$spp_catch / d$hours_fished
d$vessel <- f(d$vessel)

m1 <- fit_cpue_index_tweedie(d, cpue ~ year_factor + vessel)
m2 <- fit_cpue_index_tweedie(d, cpue ~ year_factor + (1 | vessel))

}
}
