% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-inla.R
\name{fit_inla}
\alias{fit_inla}
\title{fit_inla()}
\usage{
fit_inla(dat, response = "present", n_knots = 50, family = "binomial",
  max_edge = c(3, 10), kmeans = TRUE, plot = FALSE, fit_model = TRUE,
  extend = list(n = 8, offset = -0.1), offset = c(1, 5), cutoff = 1,
  include_depth = TRUE, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{dat}{TODO}

\item{response}{TODO}

\item{n_knots}{TODO}

\item{family}{TODO}

\item{max_edge}{TODO}

\item{kmeans}{TODO}

\item{plot}{TODO}

\item{fit_model}{TODO}

\item{extend}{TODO}

\item{offset}{TODO}

\item{cutoff}{TODO}

\item{include_depth}{TODO}
}
\description{
fit_inla()
}
