% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{format_iphc_longest}
\alias{format_iphc_longest}
\title{Format the longest IPHC time series index to agree with other surveys}
\usage{
format_iphc_longest(ser_longest)
}
\arguments{
\item{ser_longest}{Output from \code{\link[=calc_iphc_full_res]{calc_iphc_full_res()}}, but only want the
ser_longest component.}
}
\value{
Renamed ser_longest, with some required columns calculated.
}
\description{
Format the longest IPHC time series index to agree with other surveys so
that \code{\link[=plot_survey_index]{plot_survey_index()}} works automatically. So the mean catch rate
gets renames as `biomass' even though it's numbers per effective skate.
\strong{NOTE} num_sets is currently hardwired as 130, Issue 45.
}
\examples{
\dontrun{
sp = "yelloweye rockfish"
set_counts <- get_all_iphc_set_counts(sp)
iphc_set_counts_sp <- calc_iphc_full_res(set_counts)
format_iphc_longest(iphc_set_counts_sp$ser_longest)
}

}
