% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_ageing_precision}
\alias{get_ageing_precision}
\title{Get PBS ageing precision data}
\usage{
get_ageing_precision(species)
}
\arguments{
\item{species}{A character vector of a species common names}
}
\description{
Get PBS ageing precision data
}
\examples{
\dontrun{
get_ageing_precision("shortraker rockfish") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision()

get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision(n = 200)
}
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}},
  \code{\link{get_cpue_index}}, \code{\link{get_cpue}},
  \code{\link{get_sample_trips}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survey}}, \code{\link{get_survsamples}}

Other ageing precision functions: \code{\link{get_cpue_index}},
  \code{\link{plot_ageing_precision}},
  \code{\link{tidy_ageing_precision}}
}
