% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_all_data}
\alias{get_all_data}
\title{Get (and cache) all PBS groundfish data for one or more species}
\usage{
get_all_data(species, path = "data-cache")
}
\arguments{
\item{species}{A character vector of species common names}

\item{path}{The folder where the cached data will be saved}
}
\description{
Get (and cache) all PBS groundfish data for one or more species
}
