% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_bio_indices}
\alias{get_bio_indices}
\title{Get biological index data}
\usage{
get_bio_indices(spp)
}
\arguments{
\item{spp}{A character vector of species common names}
}
\description{
Get biological index data
}
