% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_commercial_specimens}
\alias{get_commercial_specimens}
\title{Get commercial specimen data}
\usage{
get_commercial_specimens(spp)
}
\arguments{
\item{spp}{A character vector of species common names}
}
\description{
Get commercial specimen data
}
