% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_cpue}
\alias{get_cpue}
\title{Get PBS commercial CPUE data (for mapping)}
\usage{
get_cpue(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS commercial CPUE data (for mapping)
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_ageing_precision}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}},
  \code{\link{get_cpue_index}},
  \code{\link{get_sample_trips}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survey}}, \code{\link{get_survsamples}}
}
