% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_cpue_historic}
\alias{get_cpue_historic}
\title{Get all fishing catch and effort to calculate historic commercial CPUE}
\usage{
get_cpue_historic(species = NULL, alt_year_start_date = "04-01",
  areas = c("3[CD]+", "5[AB]+", "5[CDE]+"), end_year = NULL)
}
\arguments{
\item{species}{Species to filter for positive fishing events. Leave as \code{NULL}
to include all fishing events. One or more species common names (e.g.
\code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species
codes can be specified as numeric vectors \code{c(396, 442}) or characters
\code{c("396", "442")}. Numeric values shorter than 3 digits will be expanded to
3 digits and converted to character objects (\code{1} turns into \code{"001"}).
Species common names and species codes should not be mixed. If any element
is missing a species code, then all elements will be assumed to be species
common names.}

\item{alt_year_start_date}{Alternative year starting date specified as a
month-day combination. E.g. "03-01" for March 1st. Can be used to create
'fishing years'.}

\item{areas}{Area groupings as a vector of regular expressions.
See \code{\link[base:regex]{base::regex()}}.}

\item{end_year}{Specify the last calendar year to be extracted.}
}
\description{
Get all fishing catch and effort to calculate historic commercial CPUE
}
