% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_landings}
\alias{get_landings}
\title{Get commercial landings data}
\usage{
get_landings(spp)
}
\arguments{
\item{spp}{A character vector of species common names}
}
\description{
Get commercial landings data
}
