% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_all_catch}
\alias{get_pbs_all_catch}
\title{Get PBS catch and effort data for all species for CPUE index calculation}
\usage{
get_pbs_all_catch(gear = "bottom trawl", min_year = 1996)
}
\arguments{
\item{gear}{A single gear type to include}

\item{min_year}{Minimum year to return}
}
\description{
Can be used to extract data for \code{\link{prep_pbs_cpue_index}}.
}
