% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_bioindex}
\alias{get_pbs_bioindex}
\title{Get PBS biological index data}
\usage{
get_pbs_bioindex(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS biological index data
}
