% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_catch}
\alias{get_pbs_catch}
\title{Get PBS commercial landings data}
\usage{
get_pbs_catch(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS commercial landings data
}
