% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_commsamples}
\alias{get_pbs_commsamples}
\title{Get PBS commercial specimen data}
\usage{
get_pbs_commsamples(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS commercial specimen data
}
