% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_cpue}
\alias{get_pbs_cpue}
\title{Get PBS commercial CPUE data}
\usage{
get_pbs_cpue(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS commercial CPUE data
}
