% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_survey}
\alias{get_pbs_survey}
\title{Get PBS spatial trawl survey data}
\usage{
get_pbs_survey(species, survey_codes = c(1, 3, 4, 16))
}
\arguments{
\item{species}{A character vector of species common names}

\item{survey_codes}{A numeric vector of survey series IDs}
}
\description{
Get PBS spatial trawl survey data
}
