% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_survsamples}
\alias{get_pbs_survsamples}
\title{Get PBS survey specimen data}
\usage{
get_pbs_survsamples(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS survey specimen data
}
