% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_sample_trips}
\alias{get_sample_trips}
\title{Title here TODO}
\usage{
get_sample_trips()
}
\description{
Title here TODO
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_ageing_precision}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}},
  \code{\link{get_cpue_index}}, \code{\link{get_cpue}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survey}}, \code{\link{get_survsamples}}
}
