% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_spatial_survey}
\alias{get_spatial_survey}
\title{Get spatial trawl survey data}
\usage{
get_spatial_survey(spp, survey_codes = c(1, 3, 4, 16))
}
\arguments{
\item{spp}{A character vector of species common names}

\item{survey_codes}{A numeric vector of survey series IDs}
}
\description{
Get spatial trawl survey data
}
