% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_survey}
\alias{get_survey}
\title{Get PBS spatial trawl survey data}
\usage{
get_survey(species, survey_codes = c(1, 3, 4, 16))
}
\arguments{
\item{species}{A character vector of species common names}

\item{survey_codes}{A numeric vector of survey series IDs}
}
\description{
Get PBS spatial trawl survey data
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_ageing_precision}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}},
  \code{\link{get_cpue_index}}, \code{\link{get_cpue}},
  \code{\link{get_sample_trips}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survsamples}}
}
