% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_survey_specimens}
\alias{get_survey_specimens}
\title{Get survey specimen data}
\usage{
get_survey_specimens(spp)
}
\arguments{
\item{spp}{A character vector of species common names}
}
\description{
Get survey specimen data
}
