% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_survsamples}
\alias{get_survsamples}
\title{Get PBS survey specimen data}
\usage{
get_survsamples(species)
}
\arguments{
\item{species}{A character vector of species common names}
}
\description{
Get PBS survey specimen data
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_ageing_precision}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}},
  \code{\link{get_cpue_index}}, \code{\link{get_cpue}},
  \code{\link{get_sample_trips}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survey}}
}
