% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfplot.R
\docType{package}
\name{gfplot}
\alias{gfplot}
\alias{gfplot-package}
\title{gfplot package}
\description{
gfplot package
}
\details{
See the README on
\href{https://github.com/seananderson/gfplot#readme}{GitHub}
}
