% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbssynopsis.R
\docType{package}
\name{gfsynopsis}
\alias{gfsynopsis}
\alias{gfsynopsis-package}
\title{gfsynopsis package}
\description{
PBS Synopsis Report package
}
\details{
See the README on
\href{https://github.com/seananderson/pbs-synopsis#readme}{GitHub}
}
