% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbll_n_grid}
\alias{hbll_n_grid}
\title{A 2km x 2km grid for the HBLL North survey}
\format{Data}
\usage{
hbll_n_grid
}
\description{
A 2km x 2km grid for the HBLL North survey
}
\keyword{datasets}
