% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbll_s_grid}
\alias{hbll_s_grid}
\title{A 2km x 2km grid for the HBLL South survey}
\format{Data}
\usage{
hbll_s_grid
}
\description{
A 2km x 2km grid for the HBLL South survey
}
\keyword{datasets}
