% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inject_filter}
\alias{inject_filter}
\title{gfplot utilities}
\usage{
inject_filter(sql_precode, species, sql_code,
  search_flag = "-- insert species here",
  conversion_func = common2codes)
}
\arguments{
\item{sql_precode}{SQL go to inject before a species list etc.}

\item{species}{A vector of species or similar objects.}

\item{sql_code}{The SQL go to operate on.}

\item{search_flag}{What to search for in terms of where to inject the code.}

\item{conversion_func}{A conversion function to apply to the species or
related vector.}
}
\description{
gfplot utilities
}
