% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{jackknife_cpue_index}
\alias{jackknife_cpue_index}
\title{Title}
\usage{
jackknife_cpue_index(object, terms = c("f(month)", "f(vessel)", "f(locality)",
  "f(depth)", "f(latitude)"))
}
\arguments{
\item{object}{TODO}

\item{terms}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

jackknife_cpue_index(m)
}
}
\seealso{
Other plotting functions: \code{\link{fit_mat_ogive}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_survey_index}}

Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{predict_cpue_index}},
  \code{\link{tidy_cpue_index}}
}
\concept{CPUE index functions}
\concept{plotting functions}
