% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-props.R
\name{join_comps_commercial}
\alias{join_comps_commercial}
\title{Join commercial catch and composition data for weighting}
\usage{
join_comps_commercial(specimen_dat, catch_dat, value, bin_size = NULL)
}
\arguments{
\item{specimen_dat}{Specimen data}

\item{catch_dat}{Catch data}

\item{value}{The unquoted column name with the values to re-weight (age or
length)}

\item{bin_size}{The binning size (if length data)}
}
\description{
Join commercial catch and composition data for weighting
}
