% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-props.R
\name{join_comps_survey}
\alias{join_comps_survey}
\title{Join survey and composition data for weighting}
\usage{
join_comps_survey(specimen_dat, survey_tows, value, bin_size = NULL)
}
\arguments{
\item{specimen_dat}{Specimen data}

\item{survey_tows}{Survey tow data}

\item{value}{The unquoted column name with the values to re-weight (age or
length)}

\item{bin_size}{The binning size (if length data)}
}
\description{
Join survey and composition data for weighting
}
