% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_assignment}
\alias{maturity_assignment}
\title{A data frame with maturity categories and assignments.}
\format{
A data frame
}
\usage{
maturity_assignment
}
\description{
A data frame with maturity categories and assignments.
}
\keyword{datasets}
