% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_short_names}
\alias{maturity_short_names}
\title{A data frame with maturity categories and names.}
\format{
A data frame
}
\usage{
maturity_short_names
}
\description{
A data frame with maturity categories and names.
}
\keyword{datasets}
