% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs2mse.R
\name{pbs2dlmtool_data}
\alias{pbs2dlmtool_data}
\title{Convert groundfish PBS data to a DLMtool data object}
\usage{
pbs2dlmtool_data(dat, name = "", area = "3[CD]+", survey = "SYN WCVI",
  max_year = 2017, min_mean_length = 10, length_bin_interval = 2)
}
\arguments{
\item{dat}{A list object output from \code{\link[=cache_pbs_data]{cache_pbs_data()}}.}

\item{name}{A name for the stock.}

\item{area}{The groundfish and statistical area to subset the catch by.}

\item{survey}{A survey abbreviation to include as the relative index of
abundance.}

\item{max_year}{The most recent year of data to include.}

\item{min_mean_length}{The minimum number of samples to include a mean length
measurement for a given year.}

\item{length_bin_interval}{An interval for the length bins.}
}
\value{
An S4 object of class DLMtool Data.
}
\description{
Takes the output from \code{\link[=cache_pbs_data]{cache_pbs_data()}} and converts it into a DLMtool data
object.
}
\examples{
\dontrun{
cache_pbs_data("pacific ocean perch", unsorted_only = FALSE)
d <- readRDS("pacific-ocean-perch.rds")
pbs2dlmtool_data(d)
}
}
