% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbs_areas}
\alias{pbs_areas}
\title{PBS groundfish areas}
\format{A data frame:
\describe{
  \item{major_stat_area_code}{major statistical area code}
  \item{major_stat_area_description}{major statistical area description}
}}
\usage{
pbs_areas
}
\description{
A dataset containing the major statistical area codes and descriptions
}
\keyword{datasets}
