% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbs_species}
\alias{pbs_species}
\title{PBS groundfish species}
\format{
A data frame:
\describe{
\item{species_code}{species code}
\item{species_scientific_name}{species scientific name}
\item{species_common_name}{species common name}
}
}
\usage{
pbs_species
}
\description{
A dataset containing species codes and names. Used in \code{\link[=tidy_cpue_index]{tidy_cpue_index()}}.
}
\keyword{datasets}
