% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-ageing-precision.R
\name{plot_ageing_precision}
\alias{plot_ageing_precision}
\title{Plot pbs ageing precision data}
\usage{
plot_ageing_precision(dat)
}
\arguments{
\item{dat}{TODO}
}
\description{
Plot pbs ageing precision data
}
\examples{
\dontrun{
d <- get_pbs_ageing_precision("pacific ocean perch")
d <- prep_pbs_ageing_precision(d)
plot_ageing_precision(d)
}
}
