% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ages.R
\name{plot_ages}
\alias{plot_ages}
\title{Plot age frequencies with bubble plots}
\usage{
plot_ages(dat, max_size = 5, sex_gap = 0.2, year_increment = 2,
  ylab = "Age (years)", year_range = NULL)
}
\arguments{
\item{dat}{A properly formatted data frame. For example, from
\code{\link{prep_pbs_ages}}. The input data frame must have the columns (in
any order): \code{year}, \code{sex} (coded as \code{"M"} and \code{"F"}),
\code{age}, \code{survey}, \code{n_scaled}.}

\item{max_size}{Maximum dot size (passed to
\code{\link[ggplot2]{scale_size_area}})}

\item{sex_gap}{Horizontal gap between male and female bubbles}

\item{year_increment}{Increment between year labels on x-axis}

\item{ylab}{Y-axis label}

\item{year_range}{TODO}
}
\description{
Plot age frequencies with bubble plots
}
\examples{
\dontrun{
d <- get_pbs_survsamples("canary rockfish")
d <- prep_pbs_ages(d)
plot_ages(d)
}
}
