% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bioindex.R
\name{plot_bioindex}
\alias{plot_bioindex}
\title{Title}
\usage{
plot_bioindex(d, col = RColorBrewer::brewer.pal(9, "Blues")[c(3, 7)],
  title = "Biomass indices")
}
\arguments{
\item{d}{TODO}

\item{col}{TODO}

\item{title}{TODO}
}
\description{
Title
}
\examples{
\dontrun{
d <- get_bio_indices("lingcod")
d <- prep_pbs_bioindex(d)
plot_bioindex(d)
}
}
