% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R, R/utils.R
\name{plot_catch}
\alias{plot_catch}
\alias{tidy_catch}
\alias{set_fishing_year}
\title{Plot catch data over time}
\usage{
tidy_catch(dat, areas = NULL, ...)

plot_catch(
  dat,
  french = FALSE,
  ylab = en2fr("Catch", french),
  xlim = c(1954, 2017),
  units = c("1000 t", "kt", "t", "kg"),
  unreliable = c(1996, 2006),
  blank_plot = FALSE
)

set_fishing_year(
  dat,
  month_fishing_starts = 1,
  day_fishing_starts = 1,
  yr_col = "year",
  date_col = "best_date",
  ...
)
}
\arguments{
\item{dat}{Input data frame. For \code{tidy_catch()} should be from \code{\link[gfdata:get_data]{gfdata::get_catch()}}.
For \code{plot_catch()} should be from \code{tidy_catch()} or be formatted similarly.
See details.}

\item{areas}{Fisheries and Oceans major statistical areas (eg. 3C, 3D, 5A,
5B, etc.). See \code{gfplot::pbs_areas} for a lookup table with descriptions.
Enter as a regular expression. For example, \code{c("5A", "5B")} or
\code{c("5[CDE]+", "5[AB]+")}.}

\item{...}{Absorb unused parameters}

\item{french}{Logical for French or English.}

\item{ylab}{Y axis label.}

\item{xlim}{X axis limits in years.}

\item{units}{A named character vector with names referring to text that will
be pasted into the y-axis label and a value defining the quantity to divide
the \code{value} column by for that unit label. Defaults to
\code{c('1000 t' = 1000000, 't' = 1000, 'kg' = 1)} if \code{units = NULL}.}

\item{unreliable}{An optional numeric vector defining years before which the
data are less reliable. Leave as \code{NA} to omit.}

\item{blank_plot}{If \code{TRUE} then a blank plot with appropriate axes will be
returned. Used in the synopsis report.}

\item{month_fishing_starts}{The month in which the fishing year starts and
ends. The \code{year} column of the returned data frame will refer to the
fishing year as determined by this and the \code{day_fishing_starts} argument.}

\item{day_fishing_starts}{The day of the month in which the fishing year
starts and ends. See \code{month_fishing_starts}}

\item{yr_col}{Name of the column in \code{dat} holding the fishing year data}

\item{date_col}{Name of the column in \code{dat} holding the date data}
}
\value{
The data frame \code{dat} with modified year data
}
\description{
Functions for plotting catch data over time as a stacked bar plot.
}
\details{
\itemize{
\item \code{tidy_catch()} Prepares PBS data for \code{plot_catch()}. This works
across one or multiple species. * \code{plot_catch()} Plots catch. The input data
frame must come from \code{tidy_catch()} or have the columns (in any order):
\code{year}, \code{gear}, \code{value} (containing catches or landings).
}
}
\examples{
\dontrun{
d <- gfdata::get_catch("lingcod")
tidy_catch(d, areas = c("5[CDE]+", "5[AB]+", "3[CD]+")) \%>\%
  plot_catch()

tidy_catch(d, areas = c("5A", "5B")) \%>\%
  plot_catch()
}
\dontrun{
d <- gfdata::get_catch("arrowtooth flounder")
d <- set_fishing_year(d, 2, 21) # Feb 21 - Feb 20 is the fishing year
}
}
