% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_coefs_cpue_index}
\alias{plot_coefs_cpue_index}
\title{Plot coefficients from a CPUE index standardization model}
\usage{
plot_coefs_cpue_index(object, coef_sub = c(depth_band = "depth", dfo_locality
  = "locality", vessel_name = "vessel", month_factor = "month", latitude_band =
  "latitude"), model_prefixes = c("Bin.", "Pos."))
}
\arguments{
\item{object}{Model output from \code{\link{fit_cpue_index}}}

\item{coef_sub}{A named character vector of any substitutions to make on the
coefficient names. Can be used to abbreviate coefficient names for the
plot.}

\item{model_prefixes}{Text to distinguish the binary and positive component
model coefficients}
}
\value{
A ggplot
}
\description{
Plot coefficients from a CPUE index standardization model
}
\details{
Note that the coefficients for predictors treated as factors (i.e.
  likely all of the predictors), the coefficients represent the difference
  from the base level factor, which would be the first factor level
  alpha-numericaly. For example, months 02 to 12 represent the estimated
  difference between that month and month 01.
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_cpue_index}},
  \code{\link{prep_pbs_cpue_index}},
  \code{\link{tidy_cpue_index}}
}
