% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_figures.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot the correlation between two variables}
\usage{
plot_cor(x, y, groups = NULL, xlab = NULL, ylab = NULL, legend.pos = "top")
}
\arguments{
\item{x, y}{Numeric vectors.}

\item{groups}{Optional grouping vector used to colour the points.}

\item{xlab, ylab}{Axis labels.}

\item{legend.pos}{Legend position.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the correlation between two variables
}
\examples{
p <- plot_cor(iris$Sepal.Length, iris$Sepal.Width)
}
