% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_cpue_index}
\alias{plot_cpue_index}
\title{Plot a delta-lognormal commercial CPUE standardization model}
\usage{
plot_cpue_index(dat)
}
\arguments{
\item{dat}{Input data frame, for example from \code{\link{tidy_cpue_index}}}
}
\value{
A ggplot object
}
\description{
Plot a delta-lognormal commercial CPUE standardization model
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_coefs_cpue_index}},
  \code{\link{prep_pbs_cpue_index}},
  \code{\link{tidy_cpue_index}}

Other plotting functions: \code{\link{plot_bioindex}},
  \code{\link{plot_catch}},
  \code{\link{plot_coefs_cpue_index}},
  \code{\link{plot_maturity}}
}
