% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_cpue_index_coefs}
\alias{plot_cpue_index_coefs}
\title{Plot coefficients from a CPUE index standardization model}
\usage{
plot_cpue_index_coefs(object, coef_sub = c(depth_band = "depth", dfo_locality
  = "locality", vessel_name = "vessel", month_factor = "month", latitude_band =
  "latitude"))
}
\arguments{
\item{object}{Model output from \code{\link{fit_cpue_index}}}

\item{coef_sub}{A named character vector of any substitutions to make on the
coefficient names. Can be used to abbreviate coefficient names for the
plot.}
}
\value{
A ggplot object
}
\description{
Plot coefficients from a CPUE index standardization model
}
\details{
Note that the coefficients for predictors treated as factors (i.e.
  likely all of the predictors), the coefficients represent the difference
  from the base level factor, which would be the first factor level
  alpha-numericaly. For example, months 02 to 12 represent the estimated
  difference between that month and month 01.
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_cpue_index}},
  \code{\link{predict_cpue_index}},
  \code{\link{tidy_pbs_cpue_index}}

Other plotting functions: \code{\link{plot_bioindex}},
  \code{\link{plot_catch}}, \code{\link{plot_cpue_index}},
  \code{\link{plot_lengths}}, \code{\link{plot_maturity}}
}
