% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{plot_lengths}
\alias{plot_lengths}
\alias{tidy_lengths_raw}
\title{Plot length frequency data}
\usage{
tidy_lengths_raw(dat, bin_size = 2, min_specimens = 20L,
  survey_series_desc = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South", "IPHC Longline Survey"),
  surveys = c("WCHG", "HS", "QCS", "WCVI", "PHMA LL (N)", "PHMA LL (S)",
  "IPHC"), year_lim = c(1996, Inf))

plot_lengths(dat, xlab = "Length (cm)", ylab = "Relative length frequency",
  fill_col = c(M = "grey80", F = "#FF000010"), line_col = c(M = "grey40", F
  = "red"), survey_col_function = NULL, alpha = 0.24)
}
\arguments{
\item{dat}{Input data frame. For \code{tidy_lengths_raw()} should be from
\code{\link[=get_survey_samples]{get_survey_samples()}} and or \code{\link[=get_comm_samples]{get_comm_samples()}}. For \code{plot_lengths()} should
be from \code{tidy_length()} or formatted similarly. See details TODO.}

\item{bin_size}{Bin size.}

\item{min_specimens}{Minimum number of specimens for histogram data to be
computed/shown.}

\item{survey_series_desc}{A character vector of survey series to include.}

\item{surveys}{A character vector of shorter/cleaner survey names to use in
the same order as \code{survey_series_desc}. These are used in the plot.}

\item{year_lim}{TODO}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{fill_col}{Fill colours for histograms. A named vector with names
\code{"F"} and \code{"M"} for female and male.}

\item{line_col}{Line colours for histograms. A named vector with names
\code{"F"} and \code{"M"} for female and male.}
}
\description{
Functions for plotting length frequency data. (Weighting not implimented
yet. TODO)
}
\details{
\itemize{
\item \code{tidy_lengths_raw()} Prepares PBS data for \code{plot_lengths()}. Works across one
or multiple species.
\item \code{plot_lengths()} Plots length frequencies for each year for selected
surveys for a single species.
}
}
\seealso{
Other age- and length-frequency functions: \code{\link{plot_ages}},
  \code{\link{weight_comps}}
}
\concept{age- and length-frequency functions}
