% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{fit_mat_ogive}
\alias{fit_mat_ogive}
\alias{plot_mat_ogive}
\title{TODO}
\usage{
fit_mat_ogive()

plot_mat_ogive(dat, type = "age")
}
\arguments{
\item{dat}{TODO}

\item{type}{TODO}
}
\description{
TODO
}
\seealso{
Other plotting functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_surv_index}}
}
\concept{plotting functions}
