% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{plot_maturity}
\alias{plot_maturity}
\title{TODO}
\usage{
plot_maturity(dat, type = "age")
}
\arguments{
\item{dat}{TODO}

\item{type}{TODO}
}
\description{
TODO
}
\seealso{
Other plotting functions: \code{\link{plot_bioindex}},
  \code{\link{plot_catch}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}}, \code{\link{plot_lengths}}
}
