% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat-freq-month.R
\name{plot_maturity_months}
\alias{plot_maturity_months}
\alias{tidy_maturity_months}
\alias{plot_maturity_months}
\title{Plot maturity frequency by month}
\usage{
tidy_maturity_months(dat, months = seq(1, 12), ssid = NULL)

plot_maturity_months(dat, max_size = 11, sex_gap = 0.2, fill_col = c(M =
  "grey50", F = "#f44256"), line_col = c(M = "grey50", F = "#f44256"),
  alpha = 0.8, title = "Maturity frequencies", n_label_pos = c(0.6, 0.8))
}
\arguments{
\item{dat}{TODO}

\item{months}{TODO}

\item{ssid}{TODO}

\item{max_size}{TODO}

\item{sex_gap}{TODO}

\item{fill_col}{TODO}

\item{line_col}{TODO}

\item{alpha}{TODO}

\item{title}{TODO}

\item{n_label_pos}{TODO}
}
\description{
Plot maturity frequency by month
}
\examples{
\dontrun{
d <- get_survey_samples("lingcod")
tidy_maturity_months(d) \%>\%
  plot_maturity_months()
}
}
