% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictor_bubbles.R
\name{plot_predictor_bubbles}
\alias{plot_predictor_bubbles}
\title{Make predictor level bubble plots}
\usage{
plot_predictor_bubbles(dat, variable, group_id = "fishing_event_id",
  ncol = 2)
}
\arguments{
\item{dat}{A data frame with the columns area, year}

\item{variable}{The column to summarize (character value)}

\item{group_id}{The events to count over (e.g. \code{"fishing_event_id"} or \code{"trip_id"})}

\item{ncol}{Number of facetted columns (facets by area)}
}
\value{
A ggplot object
}
\description{
Make predictor level bubble plots
}
\examples{
d <- data.frame(year = rep(1991:2010, each = 20),
  spp_catch = rbinom(20 * 20, size = 1, prob = 0.8),
  fishing_event_id = rep(1:20, 20),
  my_predictor = sample(gl(20, 20), size = 400),
  area = rep("a", 400))
plot_predictor_bubbles(d, "my_predictor")
}
