% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot sample availability}
\usage{
plot_samples(dat)
}
\arguments{
\item{dat}{A properly formatted data frame. For example, from
\code{\link{prep_pbs_samples}}. The input data frame must have the columns
(in any order): \code{year}, \code{type} (with the types of samples to
plot, e.g. maturity, weight), and \code{n} (with the numbers of samples).
The axis labels will be derived from \code{type} after capitalizing the
first letter.}
}
\description{
Plot sample availability
}
\examples{
d <- expand.grid(year = 1996:2016,
  type = c("maturity", "weight", "length", "age"), stringsAsFactors = FALSE)
d$n <- round(runif(nrow(d), 0, 800))
d$n[10] <- 0 # example zero
plot_samples(d)

\dontrun{
x <- prep_pbs_samples(year_range = c(1996, 2016))
x <- dplyr::filter(x, species_common_name == "abyssal skate")
plot_samples(x)
}
}
