% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-index.R
\name{plot_survey_index}
\alias{plot_survey_index}
\alias{tidy_survey_index}
\title{Plot survey relative biomass mass index}
\usage{
tidy_survey_index(dat, survey = c("SYN WCHG", "SYN HS", "SYN QCS", "SYN WCVI",
  "HBLL OUT N", "HBLL OUT S", "HBLL INS N", "HBLL INS S", "MSA HS",
  "IPHC FISS"), min_years = 3, year_range = NULL)

plot_survey_index(dat, col = brewer.pal(9, "Greys")[c(3, 7)],
  title = "Biomass indices", max_cv = 0.4, max_set_fraction = 0.05,
  survey_cols = NULL)
}
\arguments{
\item{dat}{TODO}

\item{survey}{TODO}

\item{min_years}{TODO}

\item{year_range}{TODO}

\item{col}{TODO}

\item{title}{TODO}

\item{survey_cols}{TODO}
}
\description{
Plot survey relative biomass mass index
}
\examples{
\dontrun{
get_survey_index("lingcod") \%>\%
  tidy_survey_index() \%>\%
  plot_survey_index()

# Or without pipes:
d <- get_survey_index("lingcod")
head(d)

d_tidy <- tidy_survey_index(d)
head(d_tidy)

plot_survey_index(d_tidy)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_age_precision}},
  \code{\link{tidy_cpue_index}}

Other plotting functions: \code{\link{fit_mat_ogive}}
}
\concept{plotting functions}
\concept{tidy data functions}
