% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{plot_survey_sets}
\alias{plot_survey_sets}
\title{Title TODO}
\usage{
plot_survey_sets(pred_dat, raw_dat, fill_column = "combined",
  fill_scale = viridis::scale_fill_viridis(trans = "sqrt", option = "C"),
  pt_col = "#FFFFFF90", pt_fill = "#FFFFFF60", pt_size_range = c(0.5, 9),
  show_legend = TRUE, extrapolate_depth = FALSE, extrapolation_buffer = 0,
  show_model_predictions = TRUE, utm_zone = 9,
  fill_label = "Predicted\\nbiomass\\ndensity (kg/m)",
  pt_label = "Tow density (kg/m)")
}
\arguments{
\item{pred_dat}{TODO}

\item{raw_dat}{TODO}

\item{fill_column}{TODO}

\item{fill_scale}{TODO}

\item{pt_col}{TODO}

\item{pt_fill}{TODO}

\item{pt_size_range}{TODO}

\item{show_legend}{TODO}

\item{extrapolate_depth}{TODO}

\item{extrapolation_buffer}{TODO}

\item{show_model_predictions}{TODO}

\item{utm_zone}{TODO}

\item{fill_label}{TODO}

\item{pt_label}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
x <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "Queen Charlotte Sound Synoptic Survey",
  iter = 600, chains = 1) # for speed in this example
plot_survey_sets(x$predictions, x$data)
}
}
\seealso{
Other spatial survey modelling functions: \code{\link{survey-spatial-modelling}}
}
