% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{plot_vb}
\alias{plot_vb}
\title{Plot growth TODO}
\usage{
plot_vb(object_female, object_male, downsample = 2000L, pt_alpha = 0.2,
  xlab = "Age (years)", ylab = "Length (cm)", seed = 42, lab_x = 0.4,
  lab_y = 0.3, lab_x_gap = 0.3, lab_y_gap = 0.08, col = c(Female =
  "red", Male = "black"))
}
\arguments{
\item{object_female}{TODO}

\item{object_male}{TODO}

\item{downsample}{TODO}

\item{pt_alpha}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{seed}{TODO}

\item{lab_x}{TODO}

\item{lab_y}{TODO}

\item{lab_x_gap}{TODO}

\item{lab_y_gap}{TODO}

\item{col}{TODO}
}
\description{
Plot growth TODO
}
\seealso{
Other growth functions: \code{\link{fit_vb}}
}
\concept{growth functions}
