% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_samples}
\alias{pop_samples}
\title{Example Pacific ocean perch biological samples}
\format{A data frame}
\usage{
pop_samples
}
\description{
1000 random samples of ages, lengths, weights, and maturities
}
\keyword{datasets}
