% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_surv}
\alias{pop_surv}
\title{Example Pacific Ocean Perch survey data from QCS in 2015}
\format{A data frame}
\usage{
pop_surv
}
\description{
Originally retrieved with \code{get_survey_sets("pacific ocean perch")}.
}
\keyword{datasets}
