% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{predict_cpue_index}
\alias{predict_cpue_index}
\title{Tidy a delta-lognormal commercial CPUE standardization model}
\usage{
predict_cpue_index(object, center = TRUE)
}
\arguments{
\item{object}{A model object from \code{\link{fit_cpue_index}}}

\item{center}{Should the index be centered by subtracting the mean in log space?}
}
\description{
Tidy a delta-lognormal commercial CPUE standardization model
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{tidy_pbs_cpue_index}}
}
