% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-inla.R
\name{predict_inla}
\alias{predict_inla}
\title{predict_inla()}
\usage{
predict_inla(obj, pred_grid, samples = 100L, include_depth = TRUE)
}
\arguments{
\item{obj}{TODO}

\item{pred_grid}{TODO}

\item{samples}{TODO}

\item{include_depth}{TODO}
}
\description{
predict_inla()
}
