% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-ageing-precision.R
\name{prep_pbs_ageing_precision}
\alias{prep_pbs_ageing_precision}
\title{Prepare PBS ageing precision data}
\usage{
prep_pbs_ageing_precision(dat)
}
\arguments{
\item{dat}{A data frame from \code{\link{get_pbs_ageing_precision}}}
}
\description{
Prepare PBS ageing precision data
}
