% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ages.R
\name{prep_pbs_ages}
\alias{prep_pbs_ages}
\title{Prepare PBS age data for \code{\link{plot_ages}}}
\usage{
prep_pbs_ages(species_common_name, path = "data-cache",
  survey_series_desc = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey"), survey = c("WCHG", "HS",
  "QCS", "WCVI"))
}
\arguments{
\item{species_common_name}{A species common name}

\item{path}{A path to cached data}

\item{survey_series_desc}{A character vector of survey series to include}

\item{survey}{A character vector of shorter/cleaner survey names to use in
the same order as \code{survey_series_desc}}
}
\value{
A data frame formatted for \code{\link{plot_ages}}
}
\description{
Prepare PBS age data for \code{\link{plot_ages}}
}
\examples{
\dontrun{
prep_pbs_ages("canary rockfish")
}
}
