% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bioindex.R
\name{prep_pbs_bioindex}
\alias{prep_pbs_bioindex}
\title{Title}
\usage{
prep_pbs_bioindex(d, species = "lingcod",
  surveys = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South",
  "IRF Longline Survey (North)", "IRF Longline Survey (South)",
  "Hecate Strait Multispecies Assemblage Survey", "IPHC Longline Survey"),
  survey_names = c("West Coast Haida Gwaii Synoptic",
  "Hecate Strait Synoptic", "Queen Charlotte Sound Synoptic",
  "West Coast Vancouver Island Synoptic",
  "PHMA Rockfish Longline - Outside North",
  "PHMA Rockfish Longline - Outside South", "IRF Longline Survey North",
  "IRF Longline Survey South", "Hecate Strait Multispecies Assemblage",
  "IPHC Longline"), year_min = 1975, min_years = 3)
}
\arguments{
\item{d}{TODO}

\item{species}{TODO}

\item{surveys}{TODO}

\item{survey_names}{TODO}

\item{year_min}{TODO}

\item{min_years}{TODO}
}
\description{
Title
}
\examples{
\dontrun{
d <- get_bio_indices("lingcod")
d <- prep_pbs_bioindex(d)
d
plot_bioindex(d)
}

}
