% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catches.R
\name{prep_pbs_catch}
\alias{prep_pbs_catch}
\title{Prepare PBS catch data for \code{\link{plot_catch}}}
\usage{
prep_pbs_catch(dat)
}
\arguments{
\item{dat}{TODO}
}
\value{
A data frame formatted for \code{\link{plot_catch}}
}
\description{
Prepare PBS catch data for \code{\link{plot_catch}}
}
\examples{
\dontrun{
d <- get_pbs_catch("lingcod")
prep_pbs_catch(d)
}
}
