% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{prep_pbs_cpue_index}
\alias{prep_pbs_cpue_index}
\title{Prepare commercial PBS CPUE data}
\usage{
prep_pbs_cpue_index(dat, species_common, year_range = c(1996, Inf),
  lat_range = c(48, Inf), min_positive_tows = 100,
  min_annual_positive_trips = 4, min_years_with_min_positive_trips = 4,
  area_grep_pattern = "5[CDE]+", lat_bands = seq(48, 59, 0.1),
  depth_bands = seq(50, 450, 25), gear = "BOTTOM TRAWL")
}
\arguments{
\item{dat}{An input data frame from \code{\link{get_pbs_cpue_index}}}

\item{species_common}{The species common name}

\item{year_range}{The range of years to include}

\item{lat_range}{The range of latitudes to include}

\item{min_positive_tows}{The minimum number of positive tows over all years}

\item{min_annual_positive_trips}{The minimum number of annual positive trips}

\item{min_years_with_min_positive_trips}{The number of years in which the
\code{min_annual_positive_trips} criteria needs to be met}

\item{area_grep_pattern}{A regular expression to extract the management areas
of interest}

\item{lat_bands}{A sequence of latitude bans}

\item{depth_bands}{A sequence of depth bands}

\item{gear}{Gear types}
}
\description{
Prepare commercial PBS CPUE data
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{tidy_cpue_index}}

Other prepare data functions: \code{\link{prep_pbs_ages}},
  \code{\link{prep_pbs_bioindex}},
  \code{\link{prep_pbs_catch}}
}
