% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-residuals.R
\name{qres_tweedie}
\alias{qres_tweedie}
\alias{qres_binomial}
\alias{qres_gamma}
\alias{qres_gaussian}
\alias{plot_qres_qq}
\alias{plot_qres_histogram}
\title{Quantile residuals for a Tweedie GLMM fit with glmmTMB}
\usage{
qres_tweedie(obj, response)

qres_binomial(obj, response)

qres_gamma(obj, response)

qres_gaussian(obj, response)

plot_qres_qq(qr, nmax = 10000)

plot_qres_histogram(qr, bins = 50, nmax = 10000)
}
\arguments{
\item{obj}{A model object from \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}.}

\item{response}{The name of the response column in the data. E.g. \code{"cpue"}.}

\item{qr}{A vector of quantile residuals from \code{\link[=qres_tweedie]{qres_tweedie()}}.}

\item{nmax}{The maximum number of residuals to plot. Very large values will
take a while to plot.}

\item{bins}{The number of histogram bins.}
}
\description{
Quantile residuals for a Tweedie GLMM fit with glmmTMB
}
\examples{
m <- glmmTMB::glmmTMB(cpue ~ 1,
  data = data.frame(cpue =
      tweedie::rtweedie(1000, mu = 2, phi = 5, power = 1.5)),
  family = glmmTMB::tweedie(link = "log"))
qr <- qres_tweedie(m, "cpue")
plot_qres_qq(qr)
plot_qres_histogram(qr)
}
