% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-map.R
\name{rotate_coords}
\alias{rotate_coords}
\title{Rotate coords}
\usage{
rotate_coords(x, y, rotation_angle, rotation_center)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{rotation_angle}{TODO}

\item{rotation_center}{TODO}
}
\description{
Rotate coords
}
\examples{
x <- c(1:100, rep(100, 100), 100:1, rep(1, 100))
y <- c(rep(1, 100), 1:100, rep(100, 100), 100:1)
plot(x, y, asp = 1)
points(50, 50, col = "red")
z <- rotate_coords(x = x, y = y, rotation_angle = 24,
  rotation_center = c(50, 50))
plot(z$x, z$y, asp = 1)
points(50, 50, col = "red")
}
