% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_sql}
\alias{run_sql}
\title{Run SQL}
\usage{
run_sql(database, query)
}
\arguments{
\item{database}{The name of the database.}

\item{query}{The query to run.}
}
\description{
Run SQL
}
\details{
If you need to use a user-password setup to access the databases, then you
will need to set the R options in your .Rprofile file: \code{pbs.uid}, \code{pbs.pwd},
\code{pbs.ip}, \code{pbs.sqldriver}. E.g. \code{options(pbs.uid="MyUserName")} The default
SQL driver will be \code{"SQL Server"} if not specified in the options. This will
probably work for most people. You might try using
\code{usethis::edit_r_profile()} if you need help finding your R profile file.
}
