% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trip_ids}
\alias{sample_trip_ids}
\title{Sample-Trip IDs for surveys (TEMPORARY TODO)}
\format{A data frame}
\usage{
sample_trip_ids
}
\description{
Sample-Trip IDs for surveys (TEMPORARY TODO)
}
\keyword{datasets}
