% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setData2013}
\alias{setData2013}
\title{Station data for 2013 IPHC survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 170 rows and 8 columns.}
\source{
Originally from file 2013.20-SetInfo.csv from a spreadsheet from the
IPHC, with further calculations done in private yeye15Reproduce repo of
Andrew Edwards (see data-raw there, will put into gfplot).
}
\usage{
setData2013
}
\description{
A dataset containing details of the stations for the 2013
IPHC survey; one set was deployed at each station.
}
\keyword{datasets}
