% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{sim_cpue}
\alias{sim_cpue}
\title{Simulate fake commercial CPUE data for examples and testing}
\usage{
sim_cpue(cv = 0.3, n_samples = 20, n_years = 15, n_vessels = 8)
}
\arguments{
\item{cv}{The CV for the positive timeseries.}

\item{n_samples}{The number of samples or fishing events per vessel per
year.}

\item{n_years}{The number of years.}

\item{n_vessels}{The number of the vessels.}
}
\description{
Simulate fake commercial CPUE data for examples and testing
}
\examples{
sim_cpue()
}
