% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-by-maturity.R
\name{split_catch_maturity}
\alias{split_catch_maturity}
\title{Split catch densities by mass ratio of maturity classes}
\usage{
split_catch_maturity(
  survey_sets,
  fish,
  split_dens_type = "density_kgpm2",
  survey = c("SYN HS", "SYN QCS"),
  years = NULL,
  cutoff_quantile = 0.9995,
  p_threshold = 0.5,
  use_median_ratio = TRUE,
  sample_id_re = TRUE,
  year_re = FALSE,
  plot = FALSE
)
}
\arguments{
\item{survey_sets}{Data from \code{\link[=get_survey_sets]{get_survey_sets()}}.}

\item{fish}{Data from \code{\link[=get_survey_samples]{get_survey_samples()}}.}

\item{split_dens_type}{Which density type will be split and saved as a column providing units for
adult_density and imm_density columns.}

\item{survey}{List of survey abbreviations.}

\item{years}{List of years. Default 'NULL' includes all years.}

\item{cutoff_quantile}{Set max cutoff quantile for mass modeled from lengths.}

\item{p_threshold}{Probability of maturity to split at. Default = 0.5. Alternatives are 0.05 or 0.95.}

\item{use_median_ratio}{If TRUE, uses median proportion mature (if FALSE uses mean proportion mature)
when catch too small to have biological samples collected.}

\item{sample_id_re}{If \code{TRUE} then the model will include random intercepts
for sample ID.}

\item{year_re}{Option to have lengths at maturity vary by year, but this requires the gfvelocities package.}

\item{plot}{Logical for whether to produce plots
(length-weight and length-at-maturity relationships).}
}
\description{
Split catch densities by mass ratio of maturity classes
}
\examples{
\dontrun{
d_survey_sets <- gfdata::get_survey_sets("pacific cod")
d_survey_samples <- gfdata::get_survey_samples("pacific cod")

d_by_maturity <- split_catch_maturity(d_survey_sets, d_survey_samples,
survey = c("SYN HS", "SYN QCS"),
years = NULL,
cutoff_quantile = 0.5,
plot = TRUE
)}
}
