% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_server_accessible}
\alias{sql_server_accessible}
\title{Is this a DFO Windows computer or is the server IP accessible?}
\usage{
sql_server_accessible()
}
\description{
Is this a DFO Windows computer or is the server IP accessible?
}
