% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_boundaries}
\alias{survey_boundaries}
\title{Survey boundaries}
\format{A data frame}
\usage{
survey_boundaries
}
\description{
Survey domain polygons for the synoptic trawl surveys. For use with
\code{\link[=plot_survey_sets]{plot_survey_sets()}}.
}
\keyword{datasets}
