% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageing-precision.R, R/plot-ageing-precision.R
\name{tidy_ageing_precision}
\alias{tidy_ageing_precision}
\alias{tidy_ageing_precision}
\title{Tidy PBS ageing precision data}
\usage{
tidy_ageing_precision(dat, ageing_method_codes = c(3, 17))

tidy_ageing_precision(dat, ageing_method_codes = c(3, 17))
}
\arguments{
\item{dat}{A data frame from \code{\link{get_ageing_precision}}}

\item{ageing_method_codes}{A numeric vector of ageing method codes to filter
on. Defaults to codes \code{3} and \code{17}, which represent otolith
'break and burn' and 'break and bake' methods.}

\item{dat}{A data frame from \code{\link{get_ageing_precision}}}

\item{ageing_method_codes}{A numeric vector of ageing method codes to filter on. Default
codes \code{3} and \code{17}, which represent otolith 'break and burn' and 'break and
bake' methods.}
}
\description{
Tidy PBS ageing precision data

Tidy PBS ageing precision data
}
\examples{
\dontrun{
get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision()
}
\dontrun{
get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision()
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_ages}},
  \code{\link{tidy_bioindex}}, \code{\link{tidy_catch}},
  \code{\link{tidy_cpue_index}}

Other tidy data functions: \code{\link{tidy_ages}},
  \code{\link{tidy_bioindex}}, \code{\link{tidy_catch}},
  \code{\link{tidy_cpue_index}}
}
