% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages.R
\name{tidy_ages}
\alias{tidy_ages}
\title{Tidy PBS age data for \code{\link{plot_ages}}}
\usage{
tidy_ages(dat,
  survey_series_desc = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South", "IPHC Longline Survey"),
  survey = c("WCHG", "HS", "QCS", "WCVI", "PHMA LL (N)", "PHMA LL (S)",
  "IPHC"), spp_cat_code = 1)
}
\arguments{
\item{dat}{TODO}

\item{survey_series_desc}{A character vector of survey series to include}

\item{survey}{A character vector of shorter/cleaner survey names to use in
the same order as \code{survey_series_desc}}

\item{spp_cat_code}{A numeric vector of species categorical codes to include
for the commercial samples. Defaults to \code{1}, which refers to unsorted
samples.}
}
\value{
A data frame formatted for \code{\link{plot_ages}}
}
\description{
Tidy PBS age data for \code{\link{plot_ages}}
}
\examples{
\dontrun{
d <- get_survsamples("lingcod")
tidy_ages(d)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_ageing_precision}},
  \code{\link{tidy_bioindex}}, \code{\link{tidy_catch}},
  \code{\link{tidy_cpue_index}}
}
\concept{tidy data functions}
