% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R, R/plot-catches.R
\name{tidy_catch}
\alias{tidy_catch}
\alias{tidy_catch}
\title{Tidy PBS catch data for \code{\link{plot_catch}}}
\usage{
tidy_catch(dat)

tidy_catch(dat)
}
\arguments{
\item{dat}{TODO}

\item{dat}{TODO}
}
\value{
A data frame formatted for \code{\link{plot_catch}}

A data frame formatted for \code{\link{plot_catch}}
}
\description{
Tidy PBS catch data for \code{\link{plot_catch}}

Tidy PBS catch data for \code{\link{plot_catch}}
}
\examples{
\dontrun{
d <- get_catch("lingcod")
tidy_catch(d)
}
\dontrun{
d <- get_catch("lingcod")
tidy_catch(d)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_ageing_precision}},
  \code{\link{tidy_ages}}, \code{\link{tidy_bioindex}},
  \code{\link{tidy_cpue_index}}

Other tidy data functions: \code{\link{tidy_ageing_precision}},
  \code{\link{tidy_ages}}, \code{\link{tidy_bioindex}},
  \code{\link{tidy_cpue_index}}
}
