% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-cpue-historical.R
\name{tidy_cpue_historic}
\alias{tidy_cpue_historic}
\title{Tidy historical CPUE data}
\usage{
tidy_cpue_historic(dat, species_common, year_range = c(1956, 1996),
  area = "5[CDE]+", depth_band_width = 25, clean_bins = TRUE,
  use_alt_year = FALSE, depth_bin_quantiles = c(0, 1),
  type = c("trip-level-data", "arithmetic-cpue"))
}
\arguments{
\item{dat}{Data from \code{\link[=get_cpue_historic]{get_cpue_historic()}}.}

\item{species_common}{A species common name in lowercase.}

\item{year_range}{A range of years to include. Can go up to current year but
note the management changes, particularly in 1996.}

\item{area}{A major statistical area as a regular expression.}

\item{depth_band_width}{The width of depth bands to bin the depths in for the
index standardization data set.}

\item{clean_bins}{Logical. Should the depth and latitude bands be rounded to
the nearest clean value as defined by \code{lat_band_width} or
\code{depth_band_width}? Internally, these use, for example:
\code{gfplot:::round_down_even(lat_range[1], lat_band_width)}.}

\item{use_alt_year}{Should the alternate year (e.g. fishing year) column
\code{alt_year} be used? If \code{FALSE} then the calendar year will be used. If this
is set to \code{TRUE} then the \code{year} column will be replaced with the
\code{alt_year} column.}

\item{depth_bin_quantiles}{Quantiles for the depth bands. Values above and
below these quantiles will be discarded.}

\item{type}{Should the trip-level data be returned (for possible future index
standardization) or should the arithmetic CPUE be returned (species
specific catch and all effort is summed each year and then divided.)}
}
\value{
A (tibble) data frame.
}
\description{
The function can create trip level historical CPUE for index standardization
or an "arithmetic" CPUE where the catch and effort are summed each year and
divided.
}
\details{
Note that fishing events prior to 1991 are 'rolled up' in the databases to
locality-specific trips. This function well aggregate fishing events into
trips in 1991 and after to create a consistent form of data.
}
\examples{
\donttest{
get_cpue_historic(end_year = 2016) \%>\%
  tidy_cpue_historic(species_common = "pacific cod", area = "5[CD]+")
}
}
