% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{tidy_cpue_index_coefs}
\alias{tidy_cpue_index_coefs}
\title{Extract coefficients from a CPUE index standardization model}
\usage{
tidy_cpue_index_coefs(object)
}
\arguments{
\item{object}{Model output from \code{\link[=fit_cpue_index]{fit_cpue_index()}}.}
}
\description{
Extract coefficients from a CPUE index standardization model
}
