% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catches.R
\name{tidy_pbs_catch}
\alias{tidy_pbs_catch}
\title{Tidy PBS catch data for \code{\link{plot_catch}}}
\usage{
tidy_pbs_catch(dat)
}
\arguments{
\item{dat}{TODO}
}
\value{
A data frame formatted for \code{\link{plot_catch}}
}
\description{
Tidy PBS catch data for \code{\link{plot_catch}}
}
\examples{
\dontrun{
d <- get_pbs_catch("lingcod")
tidy_pbs_catch(d)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_pbs_ages}},
  \code{\link{tidy_pbs_bioindex}},
  \code{\link{tidy_pbs_cpue_index}}
}
