% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{tidy_survey_sets}
\alias{tidy_survey_sets}
\title{Tidy the survey set data for use in modeling}
\usage{
tidy_survey_sets(
  dat,
  survey,
  years,
  utm_zone = 9,
  density_column = "density_kgpm2"
)
}
\arguments{
\item{dat}{Output from \code{\link[=get_survey_sets]{get_survey_sets()}}.}

\item{survey}{The name of a survey (see get_ssids()).}

\item{years}{The years.}

\item{utm_zone}{UTM zone.}

\item{density_column}{Name of the density column.}
}
\description{
Tidy the survey set data for use in modeling
}
