% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_samples.R
\name{bind_samples}
\alias{bind_samples}
\title{Join survey and commercial samples}
\usage{
bind_samples(dat_comm, dat_survey)
}
\arguments{
\item{dat_comm}{TODO}

\item{dat_survey}{TODO}
}
\description{
Join survey and commercial samples
}
