% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{f}
\alias{f}
\title{Force factors to be sequential within the positive or binary data sets for TMB}
\usage{
f(x)
}
\arguments{
\item{x}{A parameter name}
}
\description{
This is necessary because sometimes, for example, a certain depth factor
level appears in the binary data set (because this depth was fished at), but
not in the positive data set (because a given species was now are caught at
that depth).
}
