% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{f}
\alias{f}
\title{Force factors to be sequential within the positive or binary data sets for TMB}
\usage{
f(x)
}
\arguments{
\item{x}{A parameter name}
}
\description{
Force factors to be sequential within the positive or binary data sets for TMB
}
