% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{fit_cpue_index}
\alias{fit_cpue_index}
\title{Fit a delta-lognormal commercial CPUE standardization model}
\usage{
fit_cpue_index(dat, formula_binomial = pos_catch ~ year_factor + f(month) +
  f(vessel) + f(locality) + f(depth) + f(latitude),
  formula_lognormal = log(spp_catch/hours_fished) ~ year_factor + f(month) +
  f(vessel) + f(locality) + f(depth) + f(latitude))
}
\arguments{
\item{dat}{A data frame from \code{\link{tidy_pbs_cpue_index}}, or a similarly
formatted data frame}

\item{formula_binomial}{Formula for the binomial model}

\item{formula_lognormal}{Formula for the lognormal model}
}
\description{
Fit a delta-lognormal commercial CPUE standardization model
}
\seealso{
Other CPUE index functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{predict_cpue_index}},
  \code{\link{tidy_pbs_cpue_index}}
}
