% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-inla.R
\name{fit_inla}
\alias{fit_inla}
\title{Fit a spatial model with INLA}
\usage{
fit_inla(dat, response = "present", n_knots = 50,
  family = "binomial", max_edge = c(20, 100), kmeans = FALSE,
  plot = FALSE, fit_model = TRUE, extend = list(n = 8, offset =
  -0.1), offset = c(5, 25), cutoff = 10, include_depth = TRUE,
  verbose = FALSE, debug = FALSE, trials = 10)
}
\arguments{
\item{dat}{A data frame from \code{scale_survey_predictors()}.}

\item{response}{A character value representing the column name for the
response variable.}

\item{n_knots}{Number of knots in the mesh. Passed to \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{family}{Family to pass on to INLA. Like we should be "binomial" or
"gamma".}

\item{max_edge}{Value to pass on to \code{\link[INLA:inla.mesh.2d]{INLA::inla.mesh.2d()}}.}

\item{kmeans}{Logical for whether to use \code{\link[stats:kmeans]{stats::kmeans()}} to calculate the
knot locations before generating the mesh with \code{\link[INLA:inla.mesh.create]{INLA::inla.mesh.create()}}
(if \code{TRUE}) or use \code{\link[INLA:inla.mesh.2d]{INLA::inla.mesh.2d()}} if \code{FALSE}.}

\item{plot}{Logical for whether the INLA mesh should be plotted.}

\item{fit_model}{Logic for whether the model should be fitted. Useful for
debugging. Will return \code{NA} in the model component if \code{FALSE}.}

\item{extend}{Value to pass on to \code{\link[INLA:inla.mesh.create]{INLA::inla.mesh.create()}}.}

\item{offset}{Value to pass on to \code{\link[INLA:inla.mesh.2d]{INLA::inla.mesh.2d()}}.}

\item{cutoff}{Value to pass on to \code{\link[INLA:inla.mesh.2d]{INLA::inla.mesh.2d()}}.}

\item{include_depth}{Logical for whether to include depth as a predictor.}

\item{verbose}{Logical for verbose output from INLA.}

\item{debug}{Logical for debugging.}

\item{trials}{Number of times to retry the model if it fails to fit.}
}
\description{
This function fits a spatial model using Gaussian random fields with INLA.
The function is not intended to be used directly by the user, but
documentation is included so that you can see the argument values that you
might pass from \code{\link[=fit_survey_sets]{fit_survey_sets()}}. It is called internally by
\code{\link[=fit_survey_sets]{fit_survey_sets()}}.
}
