% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{fit_length_weight}
\alias{fit_length_weight}
\title{Fit length-weight TODO}
\usage{
fit_length_weight(dat, sex = c("female", "male"), downsample = Inf,
  min_samples = 50L, method = c("rlm", "lm"), scale_weight = 1/1000)
}
\arguments{
\item{dat}{TODO}

\item{sex}{TODO}

\item{downsample}{TODO}

\item{min_samples}{TODO}

\item{method}{TODO}

\item{scale_weight}{TODO}
}
\description{
For use with data for a single species.
}
\examples{
model_f <- fit_length_weight(pop_samples)
model_f$model
model_f$predictions
model_f$pars
model_f$data

model_m <- fit_length_weight(pop_samples, sex = "male")
plot_length_weight(model_f, model_m)
}
\seealso{
Other growth functions: \code{\link{fit_vb}},
  \code{\link{plot_growth}}
}
