% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{fit_length_weight}
\alias{fit_length_weight}
\title{Fit a length-weight model}
\usage{
fit_length_weight(dat, sex = c("female", "male", "all"),
  downsample = Inf, min_samples = 50L, method = c("rlm", "lm"),
  too_high_quantile = 1, scale_weight = 1/1000)
}
\arguments{
\item{dat}{Input data frame. Should be from \code{\link[=get_survey_samples]{get_survey_samples()}} or
\code{\link[=get_commercial_samples]{get_commercial_samples()}}.}

\item{sex}{Either "male" or "female".}

\item{downsample}{If not \code{Inf} this represents a number of fish specimens to
sample prior to model fitting. Can be useful for large data sets that you
want to fit with MCMC for testing purposes.}

\item{min_samples}{The minimum number of fish before a model will be fit.}

\item{method}{\code{"rlm"} for \code{\link[MASS:rlm]{MASS::rlm()}} or \code{"lm"} for \code{\link[stats:lm]{stats::lm()}}.}

\item{too_high_quantile}{A quantile above which to discard weights and
lengths. Can be useful for outliers. Defaults to including all data.}

\item{scale_weight}{A value to multiply all weights by. Useful for changing
units.}
}
\description{
For use with data for a single species.
}
\examples{
\dontrun{
d <- get_survey_samples("pacific ocean perch")
model_f <- fit_length_weight(d, sex = "female")
model_m <- fit_length_weight(d, sex = "male")
plot_length_weight(object_female = model_f, object_male = model_m)
}
}
