% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{fit_length_weight}
\alias{fit_length_weight}
\title{Fit length-weight TODO}
\usage{
fit_length_weight(dat, sex = c("female", "male"), downsample = Inf,
  min_samples = 50L, method = c("rlm", "lm"), scale_weight = 1/1000)
}
\arguments{
\item{dat}{TODO}

\item{sex}{TODO}

\item{downsample}{TODO}

\item{min_samples}{TODO}

\item{method}{TODO}

\item{scale_weight}{TODO}
}
\description{
For use with data for a single species.
}
\seealso{
Other growth functions: \code{\link{fit_vb}},
  \code{\link{plot_growth}}
}
\concept{growth functions}
