% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{fit_vb}
\alias{fit_vb}
\title{Fit growth TODO}
\usage{
fit_vb(dat, sex = c("female", "male"), method = c("mpd", "mcmc"),
  downsample = Inf, chains = 4L, iter = 1000L,
  cores = parallel::detectCores(), allow_slow_mcmc = FALSE,
  est_method = median, min_samples = 100L, ...)
}
\arguments{
\item{dat}{TODO}

\item{sex}{TODO}

\item{method}{TODO}

\item{downsample}{TODO}

\item{chains}{TODO}

\item{iter}{TODO}

\item{cores}{TODO}

\item{allow_slow_mcmc}{TODO}

\item{est_method}{TODO}

\item{min_samples}{TODO}

\item{...}{TODO}
}
\description{
For use with data for a single species.
}
\seealso{
Other growth functions: \code{\link{plot_vb}}
}
\concept{growth functions}
