% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{fit_vb}
\alias{fit_vb}
\title{Fit growth TODO}
\usage{
fit_vb(dat, sex = c("female", "male"), method = c("mpd", "mcmc"),
  downsample = Inf, chains = 4L, iter = 1000L,
  cores = parallel::detectCores(), allow_slow_mcmc = FALSE,
  est_method = median, min_samples = 50L, ageing_method = c(3, 17), ...)
}
\arguments{
\item{dat}{TODO}

\item{sex}{TODO}

\item{method}{TODO}

\item{downsample}{TODO}

\item{chains}{TODO}

\item{iter}{TODO}

\item{cores}{TODO}

\item{allow_slow_mcmc}{TODO}

\item{est_method}{TODO}

\item{min_samples}{TODO}

\item{...}{TODO}
}
\description{
For use with data for a single species.
}
\examples{
\dontrun{
## with `rstan::optimizing()` for the mode of the posterior density:
x <- fit_vb(pop_samples, method = "mpd")
x$model

## with MCMC via Stan:
x <- fit_vb(pop_samples, method = "mcmc")
x$pars
x$predictions
x$data
x$model
posterior <- rstan::extract(x$model)
hist(posterior$linf)

model_f <- fit_vb(pop_samples, sex = "female")
model_m <- fit_vb(pop_samples, sex = "male")
plot_vb(model_f, model_m)

## If less than `min_samples`, fit_vb() returns an empty object that
## plot_vb() will correctly parse and produce an empty plot:
obj <- fit_vb(pop_samples[1:2,])
plot_vb(obj, obj)
}
}
\seealso{
Other growth functions: \code{\link{fit_length_wt}},
  \code{\link{plot_growth}}
}
\concept{growth functions}
