% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-iphc-ser.R
\name{format_final_year_for_map}
\alias{format_final_year_for_map}
\title{Format the final year of IPHC data to use in the spatial map}
\usage{
format_final_year_for_map(set_counts, final_year)
}
\arguments{
\item{set_counts}{species-specific set-level data from \code{\link[=tidy_iphc_survey]{tidy_iphc_survey()}}}

\item{final_year}{year for which to plot the set-by-set catch rates}
}
\value{
tibble in the format required by the plotting function, with a row
of NA's if the data are not present.
}
\description{
Takes the final year of data to get into the correct format to fit into
plotting functions for gfsynopsis report.
}
\examples{
\dontrun{
# If already loaded data via gfsynopsis then just, for any species,
format_final_year_for_map(
  readRDS("../gfsynopsis/report/data-cache/iphc/china-rockfish.rds")$set_counts)

# Else to load from scratch:
set_counts <- get_all_iphc_set_counts("yelloweye rockfish")
format_final_year_for_map(set_counts, 2017)
}
}
