% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get}
\alias{get}
\alias{get_ssids}
\alias{get_surv_tows}
\alias{get_surv_samples}
\alias{get_comm_samples}
\alias{get_catch}
\alias{get_cpue_spatial}
\alias{get_cpue_index}
\alias{get_age_precision}
\alias{get_surv_index}
\alias{get_sara_dat}
\alias{cache_pbs_data}
\title{Get PBS data}
\usage{
get_ssids()

get_surv_tows(species, ssid = c(1, 3, 4, 16))

get_surv_samples(species, ssid = NULL, remove_bad_data = TRUE)

get_comm_samples(species)

get_catch(species)

get_cpue_spatial(species)

get_cpue_index(gear = "bottom trawl", min_year = 1996)

get_age_precision(species)

get_surv_index(species, ssid = NULL)

get_sara_dat()

cache_pbs_data(species, path = "data-cache")
}
\arguments{
\item{species}{One or more species common names (e.g. \code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species codes can be
specified as numeric vectors \code{c(396, 442}) or characters \code{c("396", "442")}.
Numeric values shorter than 3 digits will be expanded to 3 digits and
converted to character objects (\code{1} turns into \code{"001}). Species common
names and species codes should not be mixed. If any element is missing a
species code, then all elements will be assumed to be species, common
names.}

\item{ssid}{A numeric vector of survey series IDs. Run \code{get_ssids()} if you
want to look up a survey series ID via a surveys series description.}

\item{remove_bad_data}{Remove known bad data?}

\item{gear}{The gear type(s) to include. Will be converted to uppercase.}

\item{min_year}{Minimum year to return.}

\item{path}{The folder where the cached data will be saved}

\item{ssid}{TODO}
}
\description{
Long description here
}
\details{
\itemize{
\item \code{get_surv_tows()} does...
\item \code{get_surv_samples()} does...
\item \code{get_comm_samples()} does...
\item \code{get_catch()} does...
\item \code{get_cpue_spatial()} does...
\item \code{get_cpue_index()} does...
\item \code{get_age_precision()} does...
\item \code{get_sara_dat()} does...
\item \code{get_surv_index()} does...
\item \code{cache_pbs_data()} does...
}
}
