% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get}
\alias{get}
\alias{get_sample_trips}
\alias{get_strata}
\alias{get_survey}
\alias{get_survsamples}
\alias{get_commsamples}
\alias{get_catch}
\alias{get_cpue}
\alias{get_cpue_index}
\alias{get_ageing_precision}
\alias{get_bioindex}
\alias{get_sara_dat}
\alias{cache_pbs_data}
\title{Get PBS data}
\usage{
get_sample_trips()

get_strata()

get_survey(species, survey_codes = c(1, 3, 4, 16))

get_survsamples(species)

get_commsamples(species)

get_catch(species)

get_cpue(species)

get_cpue_index(gear = "bottom trawl", min_year = 1996)

get_ageing_precision(species)

get_bioindex(species)

get_sara_dat()

cache_pbs_data(species, path = "data-cache")
}
\arguments{
\item{species}{A character vector of species common names}

\item{survey_codes}{A numeric vector of survey series IDs}

\item{gear}{A single gear type to include}

\item{min_year}{Minimum year to return}

\item{path}{The folder where the cached data will be saved}
}
\description{
Long description here
}
\details{
\itemize{
\item \code{get_sample_trips()} does...
\item \code{get_strata()} does...
\item \code{get_survey()} does...
\item \code{get_survsamples()} does...
\item \code{get_commsamples()} does...
\item \code{get_catch()} does...
\item \code{get_cpue()} does...
\item \code{get_cpue_index()} does...
\item \code{get_ageing_precision()} does...
\item \code{get_sara_dat()} does...
\item \code{get_bioindex()} does...
\item \code{cache_pbs_data()} does...
}
}
