% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get}
\alias{get}
\alias{get_ssids}
\alias{get_major_areas}
\alias{get_age_methods}
\alias{get_survey_sets}
\alias{get_survey_samples}
\alias{get_comm_samples}
\alias{get_catch}
\alias{get_cpue_spatial}
\alias{get_cpue_spatial_ll}
\alias{get_cpue_index}
\alias{get_age_precision}
\alias{get_survey_index}
\alias{get_sara_dat}
\alias{cache_pbs_data}
\title{Get PBS data}
\usage{
get_ssids()

get_major_areas()

get_age_methods()

get_survey_sets(species, ssid = c(1, 3, 4, 16, 2, 14, 22, 36),
  join_sample_ids = FALSE)

get_survey_samples(species, ssid = NULL, remove_bad_data = TRUE)

get_comm_samples(species, discard_keepers = TRUE)

get_catch(species)

get_cpue_spatial(species)

get_cpue_spatial_ll(species)

get_cpue_index(gear = "bottom trawl", min_year = 1996)

get_age_precision(species)

get_survey_index(species, ssid = NULL)

get_sara_dat()

cache_pbs_data(species, path = "data-cache", compress = FALSE,
  min_cpue_year = 1996)
}
\arguments{
\item{species}{One or more species common names (e.g. \code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species codes can be
specified as numeric vectors \code{c(396, 442}) or characters \code{c("396", "442")}.
Numeric values shorter than 3 digits will be expanded to 3 digits and
converted to character objects (\code{1} turns into \code{"001"}). Species common
names and species codes should not be mixed. If any element is missing a
species code, then all elements will be assumed to be species common
names.}

\item{ssid}{A numeric vector of survey series IDs. Run
\code{\link{get_ssids}} if you for a look-up table of available survey
series IDs with surveys series descriptions.}

\item{join_sample_ids}{TODO}

\item{remove_bad_data}{Remove known bad data, such as unrealistic
length or weight values.}

\item{discard_keepers}{TODO}

\item{gear}{The gear type(s) to include. Will be converted to uppercase.}

\item{min_year}{Minimum year to return.}

\item{path}{The folder where the cached data will be saved}

\item{compress}{Compress the \code{.rds} file? Defaults to \code{FALSE} for faster
reading and writing at the expense of disk space.}

\item{min_cpue_year}{Minimum year for the CPUE data.}
}
\description{
Automates fisheries and research survey data extraction from DFO Pacific
groundfish databases. The output datasets feed into other functions (eg.
tidy, plot, modelling) for data visualization, which can be used as
prodcuts themselves or can be fed into automated DFO Pacific groundfish
data synopsis report production.
}
\details{
\itemize{
\item \code{get_ssids()} produces a lookup table for survey series IDs and
descriptions
\item \code{get_age_methods()} produces a lookup table for ageing method codes
and descriptions
\item \code{get_sample_trips()}produces a lookup table for sample ID and
fishing event ID
\item \code{get_strata_areas()} produces a lookup table for surveyed area for
each stratum within surveys
\item \code{get_survey_ids()} produces lookup table for survey IDs for a given
survey series ID
\item \code{get_survey_sets()} extracts survey catch data and spatial data for
plotting survey catchs on a map of British Columbia.
\item \code{get_survey_samples()} extracts all biological sample specimen records
from research surveys for given species and survey series IDs from GFBio
\item \code{get_comm_samples()} extracts all biological sample specimen records
from commercial data for given species from GFBio
\item \code{get_catch()} extracts all landing and discard records for a given species
from GFFOS.GF_MERGED_CATCH
\item \code{get_cpue_spatial()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the groundfish trawl fishery
\item \code{get_cpue_spatial_ll()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the longline fishery
\item \code{get_cpue_index()} extracts catch and effort data from
GFFOS.GF_MERGED_CATCH for the groundfish trawl fishery
\item \code{get_age_precision()} extracts age readings from biological samples for a
given species where there is a second ('precision') age reading
\item \code{get_sara_dat()} scrubs Species At Risk website for up-to-date species
status and listings
\item \code{get_survey_index()} extracts survey catch data for given species
and survey series id's
\item \code{cache_pbs_data()} runs all 'get' functions in the gfplot package
and caches extracted data to a given folder
}
}
\section{Note}{

\code{get_*} functions only extract data when performed on a computer connected to
the Pacific Biological Station DFO network. The \code{get_*} functions attempt to
detect this and will not run unless "PBS" is in \code{Sys.info()["nodename"]} and
the computer is a Windows computer.
}

\examples{
\dontrun{
get_survey_sets(species = "lingcod", ssid = 1)

get_survey_samples(species = 442, ssid = c(1,3,4,16))
get_comm_samples(c(442, 397))

species <- c("lingcod", "pacific cod")
get_catch(species)

get_cpue_spatial(species)
get_cpue_spatial_ll("yelloweye rockfish")

get_cpue_index(gear = "bottom trawl", min_year = 2012, max_year = 2017)

get_cpue_spatial("pacific cod")
get_cpue_spatial_ll("arrowtooth flounder")
}

}
\concept{get data functions}
