% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_cpue_index}
\alias{get_cpue_index}
\title{Get PBS catch and effort data for all species for CPUE index calculation}
\usage{
get_cpue_index(gear = "bottom trawl", min_year = 1996)
}
\arguments{
\item{gear}{A single gear type to include}

\item{min_year}{Minimum year to return}
}
\description{
Can be used to extract data for \code{\link{tidy_cpue_index}}.
}
\examples{
\dontrun{
get_ageing_precision("shortraker rockfish") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision()

get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision(n = 200)
}
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_ageing_precision}},
  \code{\link{get_bioindex}}, \code{\link{get_catch}},
  \code{\link{get_commsamples}}, \code{\link{get_cpue}},
  \code{\link{get_sample_trips}},
  \code{\link{get_sara_dat}}, \code{\link{get_strata}},
  \code{\link{get_survey}}, \code{\link{get_survsamples}}

Other ageing precision functions: \code{\link{get_ageing_precision}},
  \code{\link{plot_ageing_precision}},
  \code{\link{tidy_ageing_precision}}
}
