% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_data}
\alias{get_data}
\alias{get_ssids}
\alias{get_major_areas}
\alias{get_gear_types}
\alias{get_age_methods}
\alias{get_survey_sets}
\alias{get_survey_samples}
\alias{get_commercial_samples}
\alias{get_catch}
\alias{get_cpue_historic}
\alias{get_cpue_spatial}
\alias{get_cpue_spatial_ll}
\alias{get_cpue_index}
\alias{get_age_precision}
\alias{get_survey_index}
\alias{get_sara_dat}
\alias{cache_pbs_data}
\title{Get PBS data}
\usage{
get_ssids()

get_major_areas()

get_gear_types()

get_age_methods()

get_survey_sets(species, ssid = c(1, 3, 4, 16, 2, 14, 22, 36),
  join_sample_ids = FALSE, verbose = FALSE)

get_survey_samples(species, ssid = NULL, remove_bad_data = TRUE,
  unsorted_only = TRUE)

get_commercial_samples(species, unsorted_only = TRUE)

get_catch(species)

get_cpue_historic(species, fishing_year = FALSE, end_year = NA)

get_cpue_spatial(species)

get_cpue_spatial_ll(species)

get_cpue_index(gear = "bottom trawl", min_cpue_year = 1996)

get_age_precision(species)

get_survey_index(species, ssid = NULL)

get_sara_dat()

cache_pbs_data(species, path = ".", compress = FALSE,
  unsorted_only = TRUE, historic_cpue = FALSE, survey_sets = FALSE,
  verbose = TRUE)
}
\arguments{
\item{species}{One or more species common names (e.g. \code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species codes can be
specified as numeric vectors \code{c(396, 442}) or characters \code{c("396", "442")}.
Numeric values shorter than 3 digits will be expanded to 3 digits and
converted to character objects (\code{1} turns into \code{"001"}). Species common
names and species codes should not be mixed. If any element is missing a
species code, then all elements will be assumed to be species common
names.}

\item{ssid}{A numeric vector of survey series IDs. Run \code{\link[=get_ssids]{get_ssids()}} for a
look-up table of available survey series IDs with surveys series
descriptions.}

\item{join_sample_ids}{If \code{TRUE} then the sample IDs will be joined in. This
may result in repeated rows of data if the same sample ID is part of
different survey stratifications.}

\item{verbose}{If \code{TRUE} then extra messages were reprinted during data
extraction. Useful to monitor progress.}

\item{remove_bad_data}{Remove known bad data, such as unrealistic
length or weight values.}

\item{unsorted_only}{Remove sorted biological data ('keepers' and 'discards'
and unknown). Default = TRUE.}

\item{fishing_year}{Specify whether fishing year should be calendar year
(\code{FALSE}) or by Pacific Cod fishing year (\code{TRUE}) as calendar year for
<1997, Apr 1-Mar 31 for Apr 1997- Mar 2008, Apr 1 2008-Feb 20 2009, and Feb
21-Feb 20 since Feb 2009.}

\item{end_year}{Specify the last year or fishing year to be extracted.}

\item{gear}{The gear type(s) to include for CPUE. Will be converted to
uppercase. Run \code{\link[=get_gear_types]{get_gear_types()}} for a look-up table of available
gear types to select from.}

\item{min_cpue_year}{Minimum year for the CPUE data.}

\item{path}{The folder where the cached data will be saved.}

\item{compress}{Compress the \code{.rds} file? Defaults to \code{FALSE} for faster
reading and writing at the expense of disk space.}

\item{historic_cpue}{Logical for whether historical CPUE should be included.}

\item{survey_sets}{Logical for whether the survey set data should be
extracted. You might set this to \code{FALSE} if you don't need these data and
you want to substantially speed up data extraction.}
}
\value{
The \code{get_*} functions return a data frame. The \code{\link[=cache_pbs_data]{cache_pbs_data()}}
function writes an \code{.rds} file to \code{path} for each specified species. A data
object for a single species is a named list object with each element
containing a data frame from a \code{get_*} function. The element name of the list
reflects the function name with the \code{get_} part removed. For example, the
output from \code{\link[=get_survey_samples]{get_survey_samples()}} is in a list element named
\code{survey_samples()}.
}
\description{
Automates fisheries and research survey data extraction from DFO Pacific
groundfish databases. The output datasets feed into other functions (\code{tidy_},
\code{plot_}, or \code{fit_} functions) for data visualization, which can be used as
products themselves or can be fed into automated DFO Pacific groundfish data
synopsis report production.
}
\details{
\itemize{
\item \code{get_ssids()} produces a lookup table for survey series IDs and
descriptions
\item \code{get_age_methods()} produces a lookup table for ageing method codes
and descriptions
\item \code{get_sample_trips()}produces a lookup table for sample ID and
fishing event ID
\item \code{get_strata_areas()} produces a lookup table for surveyed area for
each stratum within surveys
\item \code{get_survey_ids()} produces lookup table for survey IDs for a given
survey series ID
\item \code{get_survey_sets()} extracts survey catch data and spatial data for
plotting survey catchs on a map of British Columbia.
\item \code{get_survey_samples()} extracts all biological sample specimen records
from research surveys for given species and survey series IDs from GFBio
\item \code{get_commercial_samples()} extracts all biological sample specimen records
from commercial data for given species from GFBio
\item \code{get_catch()} extracts all landing and discard records for a given species
from GFFOS.GF_MERGED_CATCH
\item \code{get_cpue_spatial()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the groundfish trawl fishery
\item \code{get_cpue_spatial_ll()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the longline fishery
\item \code{get_cpue_index()} extracts catch and effort data from
GFFOS.GF_MERGED_CATCH for the groundfish trawl fishery since 1996.
\item \code{get_cpue_historic()} extracts historical catch and effort data back into the 1950s.
\item \code{get_age_precision()} extracts age readings from biological samples for a
given species where there is a second ('precision') age reading
\item \code{get_sara_dat()} scrubs Species At Risk website for up-to-date species
status and listings
\item \code{get_survey_index()} extracts survey catch data for given species
and survey series IDs
\item \code{cache_pbs_data()} runs all 'get' functions in the gfplot package
and caches extracted data to a given folder
}

This \code{\link[=cache_pbs_data]{cache_pbs_data()}} function caches data from
\itemize{
\item \code{\link[=get_survey_samples]{get_survey_samples()}}
\item \code{\link[=get_commercial_samples]{get_commercial_samples()}}
\item \code{\link[=get_catch]{get_catch()}}
\item \code{\link[=get_cpue_spatial]{get_cpue_spatial()}}
\item \code{\link[=get_cpue_spatial_ll]{get_cpue_spatial_ll()}}
\item \code{\link[=get_survey_index]{get_survey_index()}}
\item \code{\link[=get_age_precision]{get_age_precision()}}
\item and optionally from \code{\link[=get_survey_sets]{get_survey_sets()}} and \code{\link[=get_cpue_historic]{get_cpue_historic()}}
}
}
\section{Note}{

\code{get_*} functions only extract data when performed on a computer connected to
the Pacific Biological Station DFO network.
}

\examples{
\dontrun{
## Import survey catch density and location data by tow or set for plotting
## Specify single or multiple species by common name or species code and
## single or multiple survey series id(s).
get_survey_sets(species = "lingcod", ssid = 1)

## Import survey or commercial biological data for various plots
(eg. length frequency, growth, age frequency, maturity, etc.)
get_survey_samples(species = 442, ssid = c(1, 3, 4, 16))

get_commercial_samples(c(442, 397))

## Import catch data by species for barcharts of landings by fishing area,
## geartype, and year.
get_catch("lingcod")

## Import spatial commercial catch per unit effort data for trawl or longline
## data by species for plotting along BC coast.
get_cpue_spatial("lingcod")
get_cpue_spatial_ll("yelloweye rockfish")

## Import catch and effort data by gear type for modelling commercial trawl
## cpue index.
get_cpue_index(gear = "bottom trawl", min_cpue_year = 2012)

## Import survey bootstrapped biomass estimates for plotting relative biomass
## indices by specified survey series.
get_survey_index("pacific cod", ssid = c(1, 3, 4, 16))
}

}
