% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_data}
\alias{get_data}
\alias{get_ssids}
\alias{get_major_areas}
\alias{get_management_areas}
\alias{get_fishery_ids}
\alias{get_species_groups}
\alias{get_gear_types}
\alias{get_age_methods}
\alias{get_species}
\alias{get_sensor_attributes}
\alias{get_other_surveys}
\alias{get_survey_sets}
\alias{get_survey_samples}
\alias{get_iphc_sets}
\alias{get_iphc_sets_info}
\alias{get_iphc_skates_info}
\alias{get_iphc_hooks}
\alias{get_commercial_samples}
\alias{get_catch}
\alias{get_hake_catch}
\alias{get_cpue_spatial}
\alias{get_cpue_spatial_ll}
\alias{get_cpue_index}
\alias{get_age_precision}
\alias{get_survey_index}
\alias{get_sensor_data_trawl}
\alias{get_sensor_data_ll_td}
\alias{get_sensor_data_ll_ctd}
\alias{get_management}
\alias{get_sara_dat}
\alias{cache_pbs_data}
\alias{cache_pbs_data_iphc}
\title{Get PBS data}
\usage{
get_ssids()

get_major_areas()

get_management_areas()

get_fishery_ids()

get_species_groups()

get_gear_types()

get_age_methods()

get_species()

get_sensor_attributes()

get_other_surveys()

get_survey_sets(species, ssid = c(1, 3, 4, 16, 2, 14, 22, 36),
  join_sample_ids = FALSE, verbose = FALSE)

get_survey_samples(species, ssid = NULL, remove_bad_data = TRUE,
  unsorted_only = TRUE, usability = NULL, inside = NULL)

get_iphc_sets(species, usability = NULL)

get_iphc_sets_info()

get_iphc_skates_info()

get_iphc_hooks(species, usability = NULL)

get_commercial_samples(species, unsorted_only = TRUE, usability = NULL)

get_catch(species)

get_hake_catch()

get_cpue_spatial(species)

get_cpue_spatial_ll(species)

get_cpue_index(gear = "bottom trawl", min_cpue_year = 1996)

get_age_precision(species, inside = NULL)

get_survey_index(species, ssid = NULL)

get_sensor_data_trawl(ssid = NULL, attribute = c("temperature",
  "depth", "dissolved oxygen", "salinity"), spread_attributes = FALSE)

get_sensor_data_ll_td(ssid = NULL, attribute = c("temperature",
  "depth"), sensor_min_max = FALSE)

get_sensor_data_ll_ctd(ssid = NULL, attribute = c("temperature",
  "depth", "dissolved oxygen", "salinity"), sensor_min_max = FALSE)

get_management(species = NULL, species_group = NULL, fishery = NULL,
  area = NULL, start_year = NULL)

get_sara_dat()

cache_pbs_data(species, file_name = NULL, path = ".",
  compress = FALSE, unsorted_only = TRUE, historic_cpue = FALSE,
  survey_sets = FALSE, verbose = TRUE)

cache_pbs_data_iphc(species, file_name = NULL, path = ".",
  compress = FALSE)
}
\arguments{
\item{species}{One or more species common names (e.g. \code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species codes can be
specified as numeric vectors \code{c(396, 442}) or characters \code{c("396", "442")}.
Numeric values shorter than 3 digits will be expanded to 3 digits and
converted to character objects (\code{1} turns into \code{"001"}). Species common
names and species codes should not be mixed. If any element is missing a
species code, then all elements will be assumed to be species common
names.}

\item{ssid}{A numeric vector of survey series IDs. Run \code{\link[=get_ssids]{get_ssids()}} for a
look-up table of available survey series IDs with surveys series
descriptions.}

\item{join_sample_ids}{If \code{TRUE} then the sample IDs will be joined in. This
may result in repeated rows of data if the same sample ID is part of
different survey stratifications.}

\item{verbose}{If \code{TRUE} then extra messages were reprinted during data
extraction. Useful to monitor progress.}

\item{remove_bad_data}{Remove known bad data, such as unrealistic
length or weight values.}

\item{unsorted_only}{Remove sorted biological data ('keepers' and 'discards'
and unknown). Default = TRUE.}

\item{usability}{A vector of usability codes to include. Defaults to all.
IPHC codes may be different to other surveys.}

\item{inside}{To select only the inside population (Strait of Georgia, area
4B only), set inside = 1. To select only the outside population, set inside
= 0.}

\item{gear}{The gear type(s) to include for CPUE. Will be converted to
uppercase. Run \code{\link[=get_gear_types]{get_gear_types()}} for a look-up table of available
gear types to select from.}

\item{min_cpue_year}{Minimum year for the CPUE data.}

\item{attribute}{A character vector of sensor attributes to filter for.
Run \code{get_sensor_attributes()} for a look-up table of available attributes.}

\item{sensor_min_max}{Allows for user to choose whether data are output in
wide format (= TRUE) with min and max values for each attribute for each
fishing event, or in long format (= FALSE) with only mean values for each
attribute and fishing event.}

\item{species_group}{Species group code(s) to include (see lookup table
\code{\link[=get_species_groups]{get_species_groups()}}). Defaults to all.}

\item{fishery}{The fishery_id code(s) (see lookup table \code{\link[=get_fishery_ids]{get_fishery_ids()}})
for fisheries to include in data extraction. Defaults to all.}

\item{area}{The fishery area(s) (see lookup table \code{\link[=get_management_areas]{get_management_areas()}})
to include in data extraction (eg. '5A'; c('3C', '3D', '5A', '5B')).}

\item{start_year}{The minimum year to include management actions.
Defaults to all.}

\item{file_name}{Optional filename(s) for the cached file. Defaults to the
same as the \code{species} argument.}

\item{path}{The folder where the cached data will be saved.}

\item{compress}{Compress the \code{.rds} file? Defaults to \code{FALSE} for faster
reading and writing at the expense of disk space.}

\item{historic_cpue}{Logical for whether historical CPUE should be included.}

\item{survey_sets}{Logical for whether the survey set data should be
extracted. You might set this to \code{FALSE} if you don't need these data and
you want to substantially speed up data extraction.}
}
\value{
The \code{get_*} functions return a data frame. The \code{\link[=cache_pbs_data]{cache_pbs_data()}}
function writes an \code{.rds} file to \code{path} for each specified species. A data
object for a single species is a named list object with each element
containing a data frame from a \code{get_*} function. The element name of the list
reflects the function name with the \code{get_} part removed. For example, the
output from \code{\link[=get_survey_samples]{get_survey_samples()}} is in a list element named
\code{survey_samples()}.

The \code{\link[=get_all_iphc_set_counts]{get_all_iphc_set_counts()}} function returns a data frame.
The \code{\link[=cache_pbs_data_iphc]{cache_pbs_data_iphc()}}
function writes an \code{.rds} file to \code{path} for each specified species. A data
object for a single species is a named list object with one element
containing the data frame from \code{\link[=get_all_iphc_set_counts]{get_all_iphc_set_counts()}}.
The element name of the list is \code{set_counts}.
}
\description{
Automates fisheries and research survey data extraction from DFO Pacific
groundfish databases. The output datasets feed into other functions (\code{tidy_},
\code{plot_}, or \code{fit_} functions) for data visualization, which can be used as
products themselves or can be fed into automated DFO Pacific groundfish data
synopsis report production.
}
\details{
\itemize{
\item \code{get_ssids()} produces a lookup table for survey series IDs and
descriptions
\item \code{get_age_methods()} produces a lookup table for ageing method codes
and descriptions
\item \code{get_sample_trips()}produces a lookup table for sample ID and
fishing event ID
\item \code{get_strata_areas()} produces a lookup table for surveyed area for
each stratum within surveys
\item \code{get_survey_ids()} produces lookup table for survey IDs for a given
survey series ID
\item \code{get_survey_sets()} extracts survey catch data and spatial data for
plotting survey catchs on a map of British Columbia.
\item \code{get_survey_samples()} extracts all biological sample specimen records
from research surveys for given species and survey series IDs from GFBio
\item \code{get_commercial_samples()} extracts all biological sample specimen records
from commercial data for given species from GFBio
\item \code{get_catch()} extracts all landing and discard records for a given species
from GFFOS.GF_MERGED_CATCH
\item \code{get_hake_catch()} extracts all landing and discard records for Pacific Hake
with some extra data used in the Hake assessment.
\item \code{get_cpue_spatial()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the groundfish trawl fishery
\item \code{get_cpue_spatial_ll()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the longline fishery
\item \code{get_cpue_index()} extracts catch and effort data from
GFFOS.GF_MERGED_CATCH for the groundfish trawl fishery since 1996.
\item \code{\link[=get_cpue_historical]{get_cpue_historical()}} extracts historical catch and effort data back into
the 1950s. It's help file is on a separate page; see the link.
\item \code{get_age_precision()} extracts age readings from biological samples for a
given species where there is a second ('precision') age reading
\item \code{get_sara_dat()} scrubs Species At Risk website for up-to-date species
status and listings
\item \code{get_survey_index()} extracts survey catch data for given species
and survey series IDs
\item \code{get_iphc_sets()} extracts IPHC survey data at the set level for given
species, from 2003 to present (excluding 2013 which is not in database)
\item \code{get_iphc_sets_info()} extracts IPHC survey data regarding each set, with no
species information, to give one unique row (with lat, lon etc.) for each
set, from 2003 to present (excluding 2013 which is not in database)
\item \code{get_iphc_skates_info()} extracts IPHC survey data regarding each skate,
with no species information, to give one unique row (with lat, lon etc.)
for each set, from 2003 to present (excluding 2013 which is not in database);
needed for the hooks per skate
\item \code{get_iphc_hooks()} extracts IPHC survey data at the hook level for given
species, from 2003 to present (excluding 2013 which is not in database).
If species is 'hook with bait' then it returns the hooks that were returned
with bait.
\item \code{cache_pbs_data()} runs all 'get' functions in the gfplot package
(except those specific to IPHC data) and caches extracted data to a given
folder
\item \code{cache_pbs_data_iphc()} runs \code{get_all_iphc_set_counts} for a given species
and caches extracted data to a give folder
}

This \code{\link[=cache_pbs_data]{cache_pbs_data()}} function caches data from
\itemize{
\item \code{\link[=get_survey_samples]{get_survey_samples()}}
\item \code{\link[=get_commercial_samples]{get_commercial_samples()}}
\item \code{\link[=get_catch]{get_catch()}}
\item \code{\link[=get_cpue_spatial]{get_cpue_spatial()}}
\item \code{\link[=get_cpue_spatial_ll]{get_cpue_spatial_ll()}}
\item \code{\link[=get_survey_index]{get_survey_index()}}
\item \code{\link[=get_age_precision]{get_age_precision()}}
\item and optionally from \code{\link[=get_survey_sets]{get_survey_sets()}} and \code{\link[=get_cpue_historic]{get_cpue_historic()}}
}

This \code{\link[=cache_pbs_data_iphc]{cache_pbs_data_iphc()}} function caches the data for the given
species from \code{\link[=get_all_iphc_set_counts]{get_all_iphc_set_counts()}}
}
\section{Authentication}{

\code{get_*} functions only extract data when performed on a computer connected to
the Pacific Biological Station DFO network. By default, the functions assume
that you are on an authorized DFO Windows computer where authentication with
the databases happens automatically. If instead, you wish to connect by
username and password, see the details section in \code{\link[=run_sql]{run_sql()}}.
}

\examples{
\dontrun{
## Import survey catch density and location data by tow or set for plotting
## Specify single or multiple species by common name or species code and
## single or multiple survey series id(s).
get_survey_sets(species = "lingcod", ssid = 1)

## Import survey or commercial biological data for various plots
(eg. length frequency, growth, age frequency, maturity, etc.)
get_survey_samples(species = 442, ssid = c(1, 3, 4, 16))

get_commercial_samples(c(442, 397))

## Import catch data by species for barcharts of landings by fishing area,
## geartype, and year.
get_catch("lingcod")

## Import spatial commercial catch per unit effort data for trawl or longline
## data by species for plotting along BC coast.
get_cpue_spatial("lingcod")
get_cpue_spatial_ll("yelloweye rockfish")

## Import catch and effort data by gear type for modelling commercial trawl
## cpue index.
get_cpue_index(gear = "bottom trawl", min_cpue_year = 2012)

## Import survey bootstrapped biomass estimates for plotting relative biomass
## indices by specified survey series.
get_survey_index("pacific cod", ssid = c(1, 3, 4, 16))
}

\dontrun{
cache_pbs_data_iphc("redbanded rockfish")
cache_pbs_data_iphc(c("redbanded rockfish",
                      "pacific ocean perch",
                      "made up species") )
}
}
