% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_ageing_precision}
\alias{get_pbs_ageing_precision}
\title{Get PBS ageing precision data}
\usage{
get_pbs_ageing_precision(species)
}
\arguments{
\item{species}{A character vector of a species common names}
}
\description{
Get PBS ageing precision data
}
