% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R
\name{get_pbs_survey}
\alias{get_pbs_survey}
\title{Get PBS spatial trawl survey data}
\usage{
get_pbs_survey(species, survey_codes = c(1, 3, 4, 16))
}
\arguments{
\item{species}{A character vector of species common names}

\item{survey_codes}{A numeric vector of survey series IDs}
}
\description{
Get PBS spatial trawl survey data
}
\seealso{
Other get PBS data functions: \code{\link{cache_pbs_data}},
  \code{\link{get_pbs_ageing_precision}},
  \code{\link{get_pbs_bioindex}},
  \code{\link{get_pbs_catch}},
  \code{\link{get_pbs_commsamples}},
  \code{\link{get_pbs_cpue_index}},
  \code{\link{get_pbs_cpue}},
  \code{\link{get_pbs_sample_trips}},
  \code{\link{get_pbs_strata}},
  \code{\link{get_pbs_survsamples}},
  \code{\link{get_sara_dat}}
}
