% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfplot-package.R
\docType{package}
\name{gfplot-package}
\alias{gfplot}
\alias{gfplot-package}
\title{gfplot: publication-ready figures for cancer bioinformatics}
\description{
The package collects plotting helpers for figures that recur in cancer
bioinformatics: survival curves, ROC and time-dependent ROC curves, risk
scores, dimensionality-reduction projections, boxplots, correlation plots,
lasso coefficient paths, enrichment plots, and immune-infiltration radar
charts.
}
\details{
Loading the package sets the global \code{ggplot2} theme to
\code{\link[cowplot:theme_cowplot]{cowplot::theme_cowplot()}} unless \code{options(gfplot.set_theme = FALSE)} is
set beforehand.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gflab/gfplot}
  \item \url{https://gaofeng21cn.github.io/gfplot/}
  \item Report bugs at \url{https://github.com/gflab/gfplot/issues}
}

}
\author{
\strong{Maintainer}: Feng Gao \email{gaofeng21cn@gmail.com}

Authors:
\itemize{
  \item Feng Gao \email{gaofeng21cn@gmail.com}
}

}
\keyword{internal}
