% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfplot.R
\docType{package}
\name{gfplot}
\alias{gfplot-package}
\alias{gfplot}
\title{gfplot package}
\description{
gfplot package
}
\details{
See the README on
\href{https://github.com/seananderson/gfplot#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pbs-assess/gfplot}
  \item Report bugs at \url{https://github.com/pbs-assess/gfplot/issues}
}

}
\author{
\strong{Maintainer}: Sean C. Anderson \email{sean.anderson@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Elise A. Keppel
  \item Andrew M. Edwards
  \item Chris Grandin
}

}
