% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbssynopsis.R
\docType{package}
\name{pbssynopsis}
\alias{pbssynopsis}
\alias{pbssynopsis-package}
\title{\pkg{pbssynopsis} package}
\description{
PBS Synopsis Report package
}
\details{
See the README on
\href{https://github.com/seananderson/pbs-synopsis#readme}{GitHub}
}
