% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageing-precision.R
\name{plot_ageing_precision}
\alias{plot_ageing_precision}
\title{Plot ageing precision data}
\usage{
plot_ageing_precision(dat, n = 250, jitter = 0.25, seed = 42)
}
\arguments{
\item{dat}{A properly formatted data frame, for example, from
\code{\link{tidy_ageing_precision}}. Should have columns named
TODO.}

\item{n}{Number of fish to sample if there are more fish than \code{n}.}

\item{jitter}{Amount to randomly jitter ages. Same jitter values are used for
the precision and primary ages.}

\item{seed}{If a numeric value, set the random seed so that the same rows
are sampled each time and the same jitter values are generated. If
\code{NULL} different fish will be sampled each time function is run.}
}
\description{
Plot ageing precision data
}
\examples{
\dontrun{
get_ageing_precision("shortraker rockfish") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision()

get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision(n = 200)
}
}
\seealso{
Other ageing precision functions: \code{\link{tidy_ageing_precision}}
}
\concept{ageing precision functions}
