% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages.R
\name{plot_ages}
\alias{plot_ages}
\alias{tidy_ages_raw}
\title{Plot age frequency data}
\usage{
tidy_ages_raw(dat,
  survey_series_desc = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South", "IPHC Longline Survey"),
  survey = c("WCHG", "HS", "QCS", "WCVI", "PHMA LL (N)", "PHMA LL (S)",
  "IPHC"), spp_cat_code = 1)

plot_ages(dat, max_size = 5, sex_gap = 0.2, year_increment = 2,
  ylab = "Age (years)", year_range = NULL, fill_col = c(M = "grey50", F =
  "#f44256"), line_col = c(M = "grey50", F = "#f44256"), survey_cols = NULL,
  alpha = 0.85)
}
\arguments{
\item{dat}{Input data frame. For \code{tidy_ages_raw()} should be from
\code{\link[=get_survey_samples]{get_survey_samples()}} and or \code{\link[=get_comm_samples]{get_comm_samples()}}. For \code{plot_ages()} should
be from \code{tidy_ages_raw()} or be formatted similarly. See details.}

\item{survey_series_desc}{A character vector of survey series to include.}

\item{survey}{A character vector of shorter/cleaner survey names to use in
the same order as \code{survey_series_desc}. These are used in the plot.}

\item{spp_cat_code}{A numeric vector of species categorical codes to include
for the commercial samples. Defaults to \code{1}, which refers to unsorted
samples.}

\item{max_size}{Maximum dot size (passed to \code{\link[ggplot2:scale_size_area]{ggplot2::scale_size_area()}}).}

\item{sex_gap}{Horizontal gap between male and female bubbles.}

\item{year_increment}{Increment between year labels on x axis.}

\item{ylab}{Y axis label.}

\item{year_range}{If not \code{NULL}, a the range of years to plot. Defaults to
all years included in original data.}
}
\description{
Functions for plotting age frequency data. (Weighting not implimented
yet. TODO)
}
\details{
\itemize{
\item \code{tidy_ages_raw()} Prepares PBS data for \code{plot_ages()}. Works across one
or multiple species.
\item \code{plot_ages()} Plots age frequencies for each year for selected
surveys for a single species. Input data frame should come from
\code{tidy_ages_raw()} or follow the following format: The input data frame must
have the columns (in any order): \code{year}, \code{sex} (coded as \code{"M"} and
\code{"F"}), \code{age}, \code{survey}, \code{n_scaled}.
}
}
\examples{
\dontrun{
d <- get_survey_samples("lingcod")
tidy_ages_raw(d)
plot_ages(d)

d <- get_survey_samples("canary rockfish")
tidy_ages_raw(d) \%>\%
  plot_ages()
}

}
\seealso{
Other age- and length-frequency functions: \code{\link{plot_lengths}},
  \code{\link{weight_comps}}
}
\concept{age- and length-frequency functions}
