% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ages.R
\name{plot_ages}
\alias{plot_ages}
\title{Plot age frequency data}
\usage{
plot_ages(dat, max_size = 5, sex_gap = 0.2, year_increment = 2,
  ylab = "Age (years)", year_range = NULL, line_col = c(M = "#666666", F =
  "#f44256"), survey_cols = NULL, alpha = 0.2, grid_col = "grey95",
  diagonal_lines = seq(-2100, -1850, 10))
}
\arguments{
\item{dat}{Input data frame. Should be from \code{\link[=tidy_ages_raw]{tidy_ages_raw()}} or
\code{\link[=tidy_ages_weighted]{tidy_ages_weighted()}} or be formatted similarly. See details.}

\item{max_size}{Maximum dot size (passed to \code{\link[ggplot2:scale_size_area]{ggplot2::scale_size_area()}}).}

\item{sex_gap}{Horizontal gap between male and female bubbles.}

\item{year_increment}{Increment between year labels on x axis.}

\item{ylab}{Y axis label.}

\item{year_range}{If not \code{NULL}, a the range of years to plot. Defaults to
all years included in original data.}

\item{line_col}{A named character vector of colors for male and females.}

\item{survey_cols}{If not \code{NULL}, a named character vector for different
colors for the various surveys.}

\item{alpha}{Transparency for the fill color.}

\item{grid_col}{Colour for the gridlines.}

\item{diagonal_lines}{A numeric a vector of years to start diagonal lines at
to help trace cohorts. Note that these are passed to
\code{\link[ggplot2:geom_abline]{ggplot2::geom_abline()}} as intercepts.}
}
\description{
Functions for plotting age frequency data.
}
\details{
\itemize{
\item \code{\link[=tidy_ages_raw]{tidy_ages_raw()}} or \code{\link[=tidy_ages_weighted]{tidy_ages_weighted()}} prepare PBS data for \code{plot_ages()}.
These work across one or multiple species.
\item \code{plot_ages()} Plots age frequencies for each year for selected surveys for
a single species. Input data frame should come from \code{\link[=tidy_ages_raw]{tidy_ages_raw()}} or
\code{\link[=tidy_ages_weighted]{tidy_ages_weighted()}} or follow the following format: The input data frame
must have the columns (in any order): \code{survey}, \code{year}, \code{sex} (coded as \code{"M"}
and \code{"F"}), \code{age}, \code{proportion}, \code{total} (for the total sample number label).
}
}
\examples{
\dontrun{
# main age/length data:
rs_comm_samples <- get_comm_samples("redstripe rockfish")
rs_survey_samples <- get_survey_samples("redstripe rockfish")
#
# for weighting:
rs_catch <- get_catch("redstripe rockfish")
rs_survey_sets <- get_survey_sets("redstripe rockfish")

# survey raw age frequencies:
tidy_ages_raw(rs_survey_samples,
  sample_type = "survey") \%>\%
  plot_ages()

# survey weighted age frequencies:
tidy_ages_weighted(rs_survey_samples,
  sample_type = "survey",
  dat_survey_sets = rs_survey_sets) \%>\%
  plot_ages()

# commercial raw age frequencies:
tidy_ages_raw(rs_comm_samples,
  sample_type = "commercial") \%>\%
  plot_ages()
}
}
