% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioindex.R
\name{plot_bioindex}
\alias{plot_bioindex}
\title{Title TODO}
\usage{
plot_bioindex(dat, col = RColorBrewer::brewer.pal(9, "Greys")[c(3, 7)],
  title = "Biomass indices")
}
\arguments{
\item{dat}{TODO}

\item{col}{TODO}

\item{title}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
get_bioindex("lingcod") \%>\%
  tidy_bioindex() \%>\%
  plot_bioindex()

# Or without pipes:
d <- get_bioindex("lingcod")
head(d)

d_tidy <- tidy_bioindex(d)
head(d_tidy)

plot_bioindex(d_tidy)
}
}
\seealso{
Other plotting functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_catch}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_lengths}}, \code{\link{plot_maturity}}

Other survey biomass index functions: \code{\link{get_bioindex}},
  \code{\link{tidy_bioindex}}
}
