% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioindex.R
\name{plot_bioindex}
\alias{plot_bioindex}
\title{Title TODO}
\usage{
plot_bioindex(dat, col = RColorBrewer::brewer.pal(9, "Greys")[c(3, 7)],
  title = "Biomass indices")
}
\arguments{
\item{dat}{TODO}

\item{col}{TODO}

\item{title}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
d <- get_pbs_bioindex("lingcod")
d <- tidy_pbs_bioindex(d)
plot_bioindex(d)
}
}
\seealso{
Other plotting functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_catch}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_lengths}}, \code{\link{plot_maturity}}
}
