% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catches.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catches}
\usage{
plot_catch(dat, ylab = "Landings", units = c(`1000 tons` = 1e+06, tons =
  1000, kg = 1), unreliable = c(1996, 2006), unreliable_alpha = 0.2)
}
\arguments{
\item{dat}{A properly formatted data frame. For example, from
\code{\link{prep_pbs_catch}}. The input data frame must have the columns
(in any order): \code{year}, \code{gear}, \code{value} (containing catches
or landings).}

\item{ylab}{Text to label y axis}

\item{units}{A named character vector with names referring to text that will
be pasted into the y-axis label and a value defining the quantity to divide
the \code{value} column by for that unit label.}

\item{unreliable}{An optional numeric vector defining years before which the
data are less reliable. Leave as \code{NA} to omit.}

\item{unreliable_alpha}{The alpha level for the optional unreliable
rectangles.}
}
\description{
Plot catches
}
\examples{
\dontrun{
d <- prep_pbs_catch("canary rockfish")
plot_catch(d)
}
}
