% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{plot_catch}
\alias{plot_catch}
\alias{tidy_catch}
\title{Plot catch data over time}
\usage{
tidy_catch(dat, areas = NULL)

plot_catch(dat, ylab = "Landings", xlim = c(1954, 2017), units = c(`1000
  tons` = 1e+06, tons = 1000, kg = 1), unreliable = c(1996, 2006),
  blank_plot = FALSE)
}
\arguments{
\item{dat}{Input data frame. For \code{tidy_catch()} should be from
\code{\link[=get_catch]{get_catch()}}. For \code{plot_catch()} should be from \code{tidy_ages_raw()} or be
formatted similarly. See details.}

\item{areas}{Fisheries and Oceans major statistical areas (eg.
3C, 3D, 5A, 5B, etc.). See gfplot::pbs_areas for a lookup table
with descriptions. Enter as a regular expression. For example, \code{c("5A", "5B")} or
\code{c("5[CDE]+", "5[AB]+")}.}

\item{ylab}{Y axis label.}

\item{units}{A named character vector with names referring to text that will
be pasted into the y-axis label and a value defining the quantity to divide
the \code{value} column by for that unit label.}

\item{unreliable}{An optional numeric vector defining years before which the
data are less reliable. Leave as \code{NA} to omit.}
}
\description{
Functions for plotting catch data over time as a stacked bar plot.
}
\details{
\itemize{
\item \code{tidy_catch()} Prepares PBS data for \code{plot_catch()}. This works across one
or multiple species.
\item \code{plot_catch()} Plots catch. The input data frame must come from
\code{tidy_catch()} or have the columns (in any order): \code{year}, \code{gear}, \code{value}
(containing catches or landings).
}
}
\examples{
\dontrun{
d <- get_catch("lingcod")
tidy_catch(d, areas = c("5[CDE]+", "5[AB]+", "3[CD]+")) \%>\%
  plot_catch()

tidy_catch(d, areas = c("5A", "5B")) \%>\%
  plot_catch()
}
}
