% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_figures.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot correlation}
\usage{
plot_cor(x, y, groups = NULL, xlab = NULL, ylab = NULL,
  legend.pos = "top")
}
\value{
a ggplot2 object of the plot
}
\description{
This function plots correlation between two variables.
}
\examples{

 library(cowplot)
 plot_cor(iris$Sepal.Length, iris$Sepal.Width)
}
