% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_cpue_index}
\alias{plot_cpue_index}
\title{Plot a delta-lognormal commercial CPUE standardization model}
\usage{
plot_cpue_index(dat, all_models = TRUE)
}
\arguments{
\item{dat}{Input data frame, for example from \code{\link{predict_cpue_index}}}

\item{all_models}{TODO}
}
\value{
A ggplot object
}
\description{
Plot a delta-lognormal commercial CPUE standardization model
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

jackknife_cpue_index(m)
}
}
\seealso{
Other plotting functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_bioindex}}, \code{\link{plot_catch}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_lengths}}, \code{\link{plot_maturity}}

Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{predict_cpue_index}},
  \code{\link{tidy_cpue_index}}
}
