% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_cpue_index_coefs}
\alias{plot_cpue_index_coefs}
\title{Plot coefficients from a CPUE index standardization model}
\usage{
plot_cpue_index_coefs(object)
}
\arguments{
\item{object}{Model output from \code{\link{fit_cpue_index}}}
}
\value{
A ggplot object
}
\description{
Plot coefficients from a CPUE index standardization model
}
\details{
Note that the coefficients for predictors treated as factors (i.e.
  likely all of the predictors), the coefficients represent the difference
  from the base level factor, which would be the first factor level
  alpha-numericaly. For example, month 02 represents the estimated
  difference between February and January.
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

jackknife_cpue_index(m)
}
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index}},
  \code{\link{predict_cpue_index}},
  \code{\link{tidy_cpue_index}}

Other plotting functions: \code{\link{jackknife_cpue_index}},
  \code{\link{plot_bioindex}}, \code{\link{plot_catch}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_lengths}}, \code{\link{plot_maturity}}
}
