% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-map.R
\name{plot_cpue_spatial}
\alias{plot_cpue_spatial}
\alias{plot_catch_spatial}
\title{Plot a map of commercial CPUE or Catch (trawl only)}
\usage{
plot_cpue_spatial(
  dat,
  start_year = 2013,
  bin_width = 7,
  n_minimum_vessels = 3,
  xlim = c(122, 890),
  ylim = c(5373, 6027),
  utm_zone = 9,
  bath = c(100, 200, 500),
  fill_scale = ggplot2::scale_fill_viridis_c(trans = "sqrt", option = "D"),
  colour_scale = ggplot2::scale_colour_viridis_c(trans = "sqrt", option = "D"),
  rotation_angle = 0,
  rotation_center = c(500, 5700),
  fill_lab = ifelse(plot_catch, en2fr("Catch (t)", translate = french),
    en2fr("CPUE (kg/hr)", translate = french)),
  show_historical = FALSE,
  return_data = FALSE,
  min_cells = 1,
  french = FALSE,
  percent_excluded_xy = NULL,
  percent_excluded_text = "Fishing events excluded due to Privacy Act",
  show_majorbound = FALSE,
  major_labels = boundary_labels(utm_zone, xmin = xlim[1]),
  plot_catch = FALSE
)

plot_catch_spatial(...)
}
\arguments{
\item{dat}{Data from \code{\link[gfdata:get_data]{gfdata::get_cpue_spatial()}}, \code{\link[gfdata:get_data]{gfdata::get_cpue_spatial_ll()}},
or \code{\link[gfdata:get_data]{gfdata::get_catch_spatial()}}}

\item{start_year}{Starting year.}

\item{bin_width}{Width of hexagons in km.}

\item{n_minimum_vessels}{Minimum number of unique vessels before a hexagon is
shown. Defaults to 3 to satisfy privacy requirements.}

\item{xlim}{X axis limits in UTM units.}

\item{ylim}{Y axis limits in UTM units.}

\item{utm_zone}{UTM zone.}

\item{bath}{A numeric vector of depths to show bathymetry countours at.}

\item{fill_scale}{A ggplot \verb{scale_fill_*} function to control colour shading.}

\item{colour_scale}{A ggplot \verb{scale_colour_*} function to control border of
hexagon colours. This should likely match \code{fill_scale}.}

\item{rotation_angle}{Angle to rotate the entire map. Used in the synopsis
report to rotate the coast 40 degrees to fit more plots on the page.}

\item{rotation_center}{The center in UTM coordinates around which to rotate
the coast if it is rotated at all.}

\item{fill_lab}{Label for the color legend.}

\item{show_historical}{Should the historical extent of fishing (before
\code{start_year}) be shown?}

\item{return_data}{Logical for whether to return the data instead of the
plot.}

\item{min_cells}{The minimum number of cells needed before the hexagons are
shown.}

\item{french}{Logical. If \code{TRUE}, create the plot in French}

\item{percent_excluded_xy}{If not \code{NULL}, should be a numeric vector of
length 2 corresponding to the x and y location (as fraction from the bottom
left) of text describing the percentage of fishing events excluded due to
the privacy rule.}

\item{percent_excluded_text}{The text to associate with the annotation
showing the percentage of fishing events excluded due to the privacy rule.}

\item{show_majorbound}{If \code{TRUE}, adds management boundaries.}

\item{major_labels}{Default uses internal function boundary_labels() to
retrieve a label dataframe. Function can be called and the output modified
outside this function in order to customize label placement
(e.g., \code{labels <- gfplot:::boundary_labels(utm_zone = 9)}
\code{labels[!(labels$label \%in\% c("4B", "5C", "5D")),]$X <- 200}).}

\item{plot_catch}{If \code{TRUE} plot catch instead of CPUE. If \code{FALSE} (default),
plot CPUE. See \code{dat} for data information.}

\item{...}{Arguments for \code{\link[=plot_cpue_spatial]{plot_cpue_spatial()}}.}
}
\description{
Plot a map of commercial CPUE or Catch (trawl only)
}
\examples{
## fake data demo:
xlim <- c(-134.1, -123.0)
ylim <- c(48.4, 54.25)
d <- dplyr::tibble(lat = runif(1000, min(ylim), max(ylim)),
  lon = runif(length(lat), min(xlim), max(xlim)),
  species_common_name = "fake species", fishing_event_id = 1,
  year = 2013, cpue = rlnorm(length(lat), log(1000), 0.6),
  vessel_registration_number = rep(seq_len(100), each = 10))
plot_cpue_spatial(d, bin_width = 15, n_minimum_vessels = 1)
}
