% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-map.R
\name{plot_cpue_spatial}
\alias{plot_cpue_spatial}
\title{Title TODO}
\usage{
plot_cpue_spatial(dat, bin_width = 7, n_minimum_vessels = 3,
  xlim_ll = c(-134.1, -123), ylim_ll = c(48.4, 54.25), utm_zone = 9,
  bath = c(100, 200, 500), fill_scale = viridis::scale_fill_viridis(trans =
  "sqrt", option = "C"), surv_cols = c(WCHG = "#6BAED6", HS = "#74C476", QCS =
  "#FB6A4A", WCVI = "#9E9AC8"), rotation_angle = 0, rotation_center = c(500,
  5700), fill_lab = "CPUE (kg/hr)")
}
\arguments{
\item{dat}{TODO}

\item{bin_width}{TODO}

\item{n_minimum_vessels}{TODO}

\item{xlim_ll}{TODO}

\item{ylim_ll}{TODO}

\item{utm_zone}{TODO}

\item{bath}{TODO}

\item{fill_scale}{TODO}

\item{surv_cols}{TODO}

\item{fill_lab}{TODO}
}
\description{
Title TODO
}
\examples{
## fake data demo:
xlim <- c(-134.1, -123.0)
ylim <- c(48.4, 54.25)
d <- dplyr::tibble(lat = runif(1000, min(ylim), max(ylim)),
  lon = runif(length(lat), min(xlim), max(xlim)),
  species_common_name = "fake species",
  cpue = rlnorm(length(lat), log(1000), 0.6),
  vessel_registration_number = rep(seq_len(100), each = 10))
plot_cpue_spatial(d, bin_width = 15, n_minimum_vessels = 1)
}
