% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{plot_growth}
\alias{plot_growth}
\alias{plot_vb}
\alias{plot_length_weight}
\title{Plot VB or length-weight fits TODO}
\usage{
plot_growth(object_female, object_male, type = c("vb", "length-weight"),
  downsample = 2000L, pt_alpha = 0.2, xlab = "Age (years)",
  ylab = "Length (cm)", seed = 42, lab_x = 0.4, lab_y = 0.3,
  lab_x_gap = 0.3, lab_y_gap = 0.06, col = c(Female = "#d80d0d", Male =
  "grey25"))

plot_vb(..., type = "vb")

plot_length_weight(..., type = "length-weight", xlab = "Length (cm)",
  ylab = "Weight (kg)", lab_x = 0.2, lab_y = 0.9, lab_x_gap = 0.35)
}
\arguments{
\item{object_female}{TODO}

\item{object_male}{TODO}

\item{type}{TODO}

\item{downsample}{TODO}

\item{pt_alpha}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{seed}{TODO}

\item{lab_x}{TODO}

\item{lab_y}{TODO}

\item{lab_x_gap}{TODO}

\item{lab_y_gap}{TODO}

\item{col}{TODO}

\item{...}{TODO}
}
\description{
Plot VB or length-weight fits TODO
}
\examples{
\dontrun{
model_f <- fit_vb(pop_samples, sex = "female")
model_m <- fit_vb(pop_samples, sex = "male")
plot_vb(model_f, model_m)
}

model_f <- fit_length_weight(pop_samples, sex = "female")
model_m <- fit_length_weight(pop_samples, sex = "male")
plot_length_weight(model_f, model_m)
}
\seealso{
Other growth functions: \code{\link{fit_length_weight}},
  \code{\link{fit_vb}}
}
\concept{growth functions}
