% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_KMCurve}
\alias{plot_KMCurve}
\title{Plot Kaplan-Meier Curve}
\usage{
plot_KMCurve(clinical, labels, limit = NULL, annot = NULL, color = NULL,
  font = "Arial", xlab = "Follow up", ylab = "Survival Probability",
  title = NULL, legend.pos = "top", risk.table = T, palette = "nature",
  anno.pos = "bottom", anno.x.shift = 0.5)
}
\arguments{
\item{clinical}{a survival object created by Surv function in survival package}

\item{labels}{a vector containing subtyping labels of patients}

\item{limit}{a numeric indicating the time limit of this KM plot}

\item{annot}{a character indicating the annotation showed up in the plot}

\item{color}{a vector containing colors used for different subtypes}

\item{font}{a character indicating the font used in the plot (default: "Arial")}

\item{xlab}{a character indicating the label of x-axis (default: "Follow up (weeks)")}

\item{ylab}{a character indicating the label of y-axis (default: "DFS (prob.)")}

\item{risk.table}{a logical indicating whether show risk table or not (default: FALSE)}

\item{lenged.pos}{a character indicating whether the legend is (default: "top")}
}
\value{
a ggplot2 object of the plot
}
\description{
This function plots Kaplan-Meier Curve for survival analysis.
}
\examples{

 library(survival)
 library(survminer)
 extrafont::loadfonts()
 data(myeloma)
 clin <- Surv(myeloma$time, myeloma$event)
 labs <- factor(myeloma$molecular_group)
 plot_KMCurve(clin, labs)
}
