% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{plot_lengths}
\alias{plot_lengths}
\title{TODO}
\usage{
plot_lengths(dat, n_bins = 25, bin_size = NULL, min_specimens = 20L)
}
\arguments{
\item{dat}{TODO}

\item{n_bins}{TODO}

\item{bin_size}{TODO}

\item{min_specimens}{TODO}
}
\description{
TODO
}
\seealso{
Other plotting functions: \code{\link{plot_bioindex}},
  \code{\link{plot_catch}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{plot_maturity}}
}
