% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lengths.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Plot length frequency data}
\usage{
plot_lengths(dat, xlab = "Length (cm)",
  ylab = "Relative length frequency", fill_col = c(M = "grey80", F =
  "#FF000010"), line_col = c(M = "grey40", F = "red"),
  survey_cols = NULL, alpha = 0.24, bin_size = 2, min_total = 20)
}
\arguments{
\item{dat}{A data frame from \code{\link[=tidy_lengths_raw]{tidy_lengths_raw()}} or \code{\link[=tidy_lengths_weighted]{tidy_lengths_weighted()}}
or a similarly formatted data frame. Must have columns named \code{survey},
\code{year}, \code{sex} (\code{"M"} and \code{"F"}), \code{length_bin}, \code{proportion}, \code{total} (for
the total number of samples for that year and survey/commercial
combination).}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{fill_col}{Fill colours for histograms. A named vector with names
\code{"F"} and \code{"M"} for female and male.}

\item{line_col}{Line colours for histograms. A named vector with names
\code{"F"} and \code{"M"} for female and male.}

\item{survey_cols}{An optional named character vector of colors for the
various surveys.}

\item{alpha}{Transparency for the fill color in the histograms.}

\item{bin_size}{Bin size. Should match the bin size used with the \code{tidy_*()}
function.}

\item{min_total}{Minimum number of fish for a given survey and year needed
before a histogram is shown.}
}
\description{
Functions for plotting length frequency data.
}
\details{
\itemize{
\item \code{\link[=tidy_lengths_raw]{tidy_lengths_raw()}} or \code{\link[=tidy_lengths_weighted]{tidy_lengths_weighted()}} prepares PBS data for
\code{plot_lengths()}. Works across one or multiple species.
\item \code{plot_lengths()} Plots length frequencies for each year for selected
surveys for a single species.
}
}
\examples{
\dontrun{
# # main age/length data:
# rs_comm_samples <- get_commercial_samples("redstripe rockfish")
# rs_survey_samples <- get_survey_samples("redstripe rockfish")
#
# # for weighting:
# rs_catch <- get_catch("redstripe rockfish")
# rs_survey_sets <- get_survey_sets("redstripe rockfish")

# survey raw length frequencies:
tidy_lengths_raw(rs_survey_samples,
  sample_type = "survey",
  bin_size = 2) \%>\%
  plot_lengths()

# survey weighted length frequencies:
tidy_lengths_weighted(rs_survey_samples,
  sample_type = "survey",
  bin_size = 2,
  dat_survey_sets = rs_survey_sets) \%>\%
  plot_lengths()

# commercial raw length frequencies:
tidy_lengths_raw(rs_comm_samples,
  sample_type = "commercial",
  bin_size = 2) \%>\%
  plot_lengths()
}
}
