% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat-freq-month.R
\name{plot_maturity_months}
\alias{plot_maturity_months}
\alias{tidy_maturity_months}
\title{Plot maturity frequency by month}
\usage{
tidy_maturity_months(
  dat,
  months = seq(1, 12),
  usability_codes = c(0, 1, 2, 6),
  french = FALSE
)

plot_maturity_months(
  dat,
  max_size = 9,
  sex_gap = 0.2,
  fill_col = c(M = "grey70", F = "black"),
  line_col = c(M = "grey70", F = "black"),
  alpha = 0.8,
  title = "Maturity frequencies",
  n_label_pos = c(0.7, 1.3),
  min_fish = 0,
  french = FALSE
)
}
\arguments{
\item{dat}{For \code{\link[=tidy_maturity_months]{tidy_maturity_months()}}, data from \code{\link[=get_survey_samples]{get_survey_samples()}}
or \code{\link[=get_commercial_samples]{get_commercial_samples()}} or \code{\link[=bind_samples]{bind_samples()}}. For \code{\link[=plot_maturity_months]{plot_maturity_months()}},
data from \code{\link[=tidy_maturity_months]{tidy_maturity_months()}}.}

\item{months}{A vector of months to include. Defaults to all.}

\item{usability_codes}{An optional vector of usability codes.
All usability codes not in this vector will be omitted.
Set to \code{NULL} to include all samples.}

\item{french}{Logical for French or English.}

\item{max_size}{The maximum size of the circles.}

\item{sex_gap}{A gap to separate the male and female circles.}

\item{fill_col}{The fill color of the circles.}

\item{line_col}{The line color of the circles.}

\item{alpha}{The transparency of the circles.}

\item{title}{A title to add to the plot.}

\item{n_label_pos}{A numeric vector of length 2 that gives the y position of
the text describing the number of male and female samples within each month
bin.}

\item{min_fish}{The minimum number of fish for the circles to be plotted for
a given month.}
}
\description{
Plot maturity frequency by month
}
\examples{
\dontrun{
# pop_samples <- get_survey_samples("pacific ocean perch")
tidy_maturity_months(pop_samples) \%>\%
  plot_maturity_months()
}
\dontrun{
d <- get_survey_samples("lingcod")
tidy_maturity_months(d) \%>\%
  plot_maturity_months()
}
}
