% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{plot_ROC}
\alias{plot_ROC}
\title{Plot ROC curve}
\usage{
plot_ROC(scores, labels, force05 = F, palette = "jama",
  legend.pos = c(0.2, 0.15), title = NULL, font = "Arial",
  percent.style = F)
}
\value{
a ggplot2 object of the plot
}
\description{
This function plots ROC curve for survival analysis.
}
\examples{
library(gfplot)
library(survminer)
data(myeloma)
extrafont::loadfonts()
plot_ROC(myeloma$CCND1, myeloma$event)
plot_ROC(myeloma[, 7:10], myeloma$event)

}
