% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{plot_sample_avail}
\alias{plot_sample_avail}
\title{Plot sample availability}
\usage{
plot_sample_avail(dat, year_range = NULL, title = "Biological samples",
  palette = "Greys", trans = sqrt)
}
\arguments{
\item{dat}{An input data frame from, for example,
\code{\link[=tidy_sample_avail]{tidy_sample_avail()}}. The input data frame must have the columns:
\describe{
\item{\code{year}}{The year.}
\item{\code{type}}{The types of samples to plot, e.g. "maturity",
"weight", "age". The axis labels will be derived from \code{type} after
capitalizing the first letter.}
\item{\code{n}}{The number of samples available for that sample type.}
}}

\item{year_range}{An optional year range.}

\item{title}{A title for the plot.}

\item{palette}{A palette to pass to \code{\link[ggplot2:scale_fill_distiller]{ggplot2::scale_fill_distiller()}}.}

\item{trans}{A function to transform the counts before applying the color
scale. E.g. \code{sqrt} or \code{log} or \code{I} if you don't want any transformation.}
}
\description{
Plot sample availability
}
\examples{
d <- expand.grid(year = 1996:2016,
  type = c("maturity", "weight", "length", "age"), stringsAsFactors = FALSE)
d$n <- round(runif(nrow(d), 0, 800))
d$n[10] <- 0 # example zero
plot_sample_avail(d)

\dontrun{
d <- get_survey_samples("lingcod")
d <- tidy_sample_avail(d, year_range = c(1996, 2016))
plot_sample_avail(d, year_range = c(1996, 2016),
  title = "Survey samples")

d <- get_comm_samples("lingcod")
d <- tidy_sample_avail(d, year_range = c(1996, 2016))
plot_sample_avail(d, year_range = c(1996, 2016),
  title = "Commercial samples")
}
}
