% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-index.R
\name{plot_survey_index}
\alias{plot_survey_index}
\alias{tidy_survey_index}
\title{Plot survey relative biomass mass index}
\usage{
tidy_survey_index(dat, surveys = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South",
  "IRF Longline Survey (North)", "IRF Longline Survey (South)",
  "Hecate Strait Multispecies Assemblage Survey", "IPHC Longline Survey"),
  survey_names = c("West Coast Haida Gwaii Synoptic",
  "Hecate Strait Synoptic", "Queen Charlotte Sound Synoptic",
  "West Coast Vancouver Island Synoptic", "PHMA Rockfish LL - North",
  "PHMA Rockfish LL - South", "IRF LL Survey North", "IRF LL Survey South",
  "Hecate Strait Multispecies Assemblage", "IPHC Longline"), min_years = 3,
  year_range = NULL)

plot_survey_index(dat, col = brewer.pal(9, "Greys")[c(3, 7)],
  title = "Biomass indices")
}
\arguments{
\item{dat}{TODO}

\item{surveys}{TODO}

\item{survey_names}{TODO}

\item{min_years}{TODO}

\item{year_range}{TODO}

\item{col}{TODO}

\item{title}{TODO}
}
\description{
Plot survey relative biomass mass index
}
\examples{
\dontrun{
get_survey_index("lingcod") \%>\%
  tidy_survey_index() \%>\%
  plot_survey_index()

# Or without pipes:
d <- get_survey_index("lingcod")
head(d)

d_tidy <- tidy_survey_index(d)
head(d_tidy)

plot_survey_index(d_tidy)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_age_precision}},
  \code{\link{tidy_cpue_index}}

Other plotting functions: \code{\link{fit_mat_ogive}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index_jk}},
  \code{\link{plot_cpue_index}}
}
\concept{plotting functions}
\concept{tidy data functions}
