% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{plot_survey_sets}
\alias{plot_survey_sets}
\title{Title TODO}
\usage{
plot_survey_sets(pred_dat, raw_dat, fill_column = "combined",
  fill_scale = viridis::scale_fill_viridis(trans = "sqrt", option = "C"),
  colour_scale = viridis::scale_colour_viridis(trans = "sqrt", option = "C"),
  pos_pt_col = "#FFFFFF60", bin_pt_col = "#FFFFFF40",
  pos_pt_fill = "#FFFFFF05", pt_size_range = c(0.5, 9),
  show_legend = TRUE, extrapolate_depth = TRUE, extrapolation_buffer = 0,
  show_model_predictions = TRUE, show_raw_data = TRUE, utm_zone = 9,
  fill_label = "Predicted\\nbiomass\\ndensity (kg/m^2)",
  pt_label = "Tow density (kg/km^2)", rotation_angle = 0,
  rotation_center = c(500, 5700), show_axes = TRUE, xlim = NULL,
  ylim = NULL, x_buffer = c(-5, 5), y_buffer = c(-5, 5),
  north_symbol = FALSE, north_symbol_coord = c(130, 5975),
  north_symbol_length = 30, cell_size = 2)
}
\arguments{
\item{pred_dat}{TODO}

\item{raw_dat}{TODO}

\item{fill_column}{TODO}

\item{fill_scale}{TODO}

\item{pt_size_range}{TODO}

\item{show_legend}{TODO}

\item{extrapolate_depth}{TODO}

\item{extrapolation_buffer}{TODO}

\item{show_model_predictions}{TODO}

\item{utm_zone}{TODO}

\item{fill_label}{TODO}

\item{pt_label}{TODO}

\item{pt_col}{TODO}

\item{pt_fill}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
x <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "SYN QCS",
  iter = 600, chains = 1) # for speed in this example
plot_survey_sets(x$predictions, x$data)
}
}
\seealso{
Other spatial survey modelling functions: \code{\link{make_prediction_grid}}
}
\concept{spatial survey modelling functions}
