% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{plot_survey_sets}
\alias{plot_survey_sets}
\title{Title TODO}
\usage{
plot_survey_sets(pred_dat, raw_dat, fill_column,
  fill_scale = viridis::scale_fill_viridis(trans = "sqrt", option = "D"),
  pt_col = "#FFFFFF90", pt_fill = "#FFFFFF60", pt_size_range = c(2, 7),
  show_legend = TRUE, extrapolate_depth = FALSE, extrapolation_buffer = 5,
  show_model_predictions = TRUE, utm_zone = 9)
}
\arguments{
\item{pred_dat}{TODO}

\item{raw_dat}{TODO}

\item{fill_column}{TODO}

\item{fill_scale}{TODO}

\item{pt_col}{TODO}

\item{pt_fill}{TODO}

\item{pt_size_range}{TODO}

\item{show_legend}{TODO}

\item{extrapolate_depth}{TODO}

\item{extrapolation_buffer}{TODO}

\item{show_model_predictions}{TODO}

\item{utm_zone}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
x <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "Queen Charlotte Sound Synoptic Survey",
  iter = 600, chains = 1, mcmc_posterior_samples = 100)
plot_survey_sets(x$predictions, x$data, fill_column = "combined")
}
}
\seealso{
Other spatial survey modelling functions: \code{\link{survey-spatial-modelling}}
}
\concept{spatial survey modelling functions}
