% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_samples}
\alias{pop_samples}
\title{Example Pacific Ocean Perch biological samples from QCS.}
\format{
A data frame
}
\usage{
pop_samples
}
\description{
Originally retrieved with \code{gfdata::get_survey_samples("pacific ocean perch")}. Has
been filtered in various ways to condense the data set for use in examples
throughout the package.
}
\keyword{datasets}
