% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{predict_cpue_index}
\alias{predict_cpue_index}
\title{Predit from a delta-lognormal commercial CPUE standardization model}
\usage{
predict_cpue_index(object, center = FALSE)
}
\arguments{
\item{object}{A model object from \code{\link{fit_cpue_index}}}

\item{center}{Should the index be centered by subtracting the mean in link space?}
}
\description{
Predit from a delta-lognormal commercial CPUE standardization model
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

jackknife_cpue_index(m)
}
}
\seealso{
Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{tidy_cpue_index}}
}
