% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ageing-precision.R
\name{prep_pbs_ageing_precision}
\alias{prep_pbs_ageing_precision}
\title{Prepare PBS ageing precision data}
\usage{
prep_pbs_ageing_precision(dat, ageing_method_codes = c(3, 17))
}
\arguments{
\item{dat}{A data frame from \code{\link{get_pbs_ageing_precision}}}

\item{ageing_method_codes}{A numeric vector of ageing method codes to filter on. Default
codes \code{3} and \code{17}, which represent otolith 'break and burn' and 'break and
bake' methods.}
}
\description{
Prepare PBS ageing precision data
}
\examples{
\dontrun{
get_pbs_ageing_precision("pacific ocean perch") \%>\%
  prep_pbs_ageing_precision()
}
}
