% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ages.R
\name{prep_pbs_ages}
\alias{prep_pbs_ages}
\title{Prepare PBS age data for \code{\link{plot_ages}}}
\usage{
prep_pbs_ages(dat,
  survey_series_desc = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South", "IPHC Longline Survey"),
  survey = c("WCHG", "HS", "QCS", "WCVI", "PHMA LL (N)", "PHMA LL (S)",
  "IPHC"))
}
\arguments{
\item{dat}{TODO}

\item{survey_series_desc}{A character vector of survey series to include}

\item{survey}{A character vector of shorter/cleaner survey names to use in
the same order as \code{survey_series_desc}}
}
\value{
A data frame formatted for \code{\link{plot_ages}}
}
\description{
Prepare PBS age data for \code{\link{plot_ages}}
}
\examples{
\dontrun{
d <- get_pbs_survsamples("lingcod")
prep_pbs_ages(d)
}
}
