% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catches.R
\name{prep_pbs_catch}
\alias{prep_pbs_catch}
\title{Prepare PBS catch data for \code{\link{plot_catch}}}
\usage{
prep_pbs_catch(species_common_name, path = "data-cache")
}
\arguments{
\item{species_common_name}{A species common name}

\item{path}{A path to cached data}
}
\value{
A data frame formatted for \code{\link{plot_catch}}
}
\description{
Prepare PBS catch data for \code{\link{plot_catch}}
}
\examples{
\dontrun{
prep_pbs_catch("canary rockfish")
}
}
