% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{prep_pbs_samples}
\alias{prep_pbs_samples}
\title{Prepare PBS samples data for \code{\link{plot_samples}}}
\usage{
prep_pbs_samples(path = "data-cache", year_range = NULL)
}
\arguments{
\item{path}{Path to cached data}

\item{year_range}{Either \code{NULL}, in which case all years are returned,
or a numeric vector of length two giving the lower and upper years to
include.}
}
\description{
Prepare PBS samples data for \code{\link{plot_samples}}
}
\examples{
\dontrun{
x <- prep_pbs_samples(year_range = c(1996, 2016))
}
}
