% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{prep_pbs_samples}
\alias{prep_pbs_samples}
\title{Prepare PBS samples data for \code{\link{plot_samples}}}
\usage{
prep_pbs_samples(dat, year_range = NULL)
}
\arguments{
\item{dat}{Input data frame from \code{\link{get_pbs_survsamples}}.}

\item{year_range}{Either \code{NULL}, in which case all years are returned,
or a numeric vector of length two giving the lower and upper years to
include.}
}
\description{
Prepare PBS samples data for \code{\link{plot_samples}}
}
\examples{
\dontrun{
d <- get_pbs_survsamples("lingcod")
prep_pbs_samples(d)

d <- get_pbs_commsamples("lingcod")
prep_pbs_samples(d)
}
}
