% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{rotate_coords}
\alias{rotate_coords}
\title{Title TODO}
\usage{
rotate_coords(x, y, rotation_angle, rotation_center)
}
\arguments{
\item{pred_dat}{TODO}

\item{raw_dat}{TODO}

\item{fill_column}{TODO}

\item{fill_scale}{TODO}

\item{pt_col}{TODO}

\item{pt_fill}{TODO}

\item{pt_size_range}{TODO}

\item{show_legend}{TODO}

\item{extrapolate_depth}{TODO}

\item{extrapolation_buffer}{TODO}

\item{show_model_predictions}{TODO}

\item{utm_zone}{TODO}

\item{fill_label}{TODO}

\item{pt_label}{TODO}
}
\description{
Title TODO
}
\examples{
\dontrun{
x <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "Queen Charlotte Sound Synoptic Survey",
  iter = 600, chains = 1) # for speed in this example
plot_survey_sets(x$predictions, x$data)
}
x <- c(1:100, rep(100, 100), 100:1, rep(1, 100))
y <- c(rep(1, 100), 1:100, rep(100, 100), 100:1)
plot(x, y, asp = 1)
points(50, 50, col = "red")
z <- rotate_coords(x = x, y = y, rotation_angle = 24,
  rotation_center = c(50, 50))
plot(z$x, z$y, asp = 1)
points(50, 50, col = "red")
}
\seealso{
Other spatial survey modelling functions: \code{\link{survey-spatial-modelling}}
}
\concept{spatial survey modelling functions}
