% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_sf}
\alias{rotate_sf}
\title{Rotate spatial features for plotting BC coastline}
\usage{
rotate_sf(sf_obj, angle = -40, lonc = -129)
}
\arguments{
\item{sf_obj}{An sf object to rotate}

\item{angle}{Rotation angle in degrees (default: -40). Negative values rotate clockwise.}

\item{lonc}{Central meridian longitude (default: -129)}
}
\value{
sf object in rotated oblique Mercator projection
}
\description{
Rotates sf objects using oblique Mercator projection to align BC's
northwest-southeast coastline with plot axes.
}
\examples{
\dontrun{
coast <- pacea::bc_coast
coast_rotated <- rotate_sf(coast, angle = -40, lonc = -129)

ggplot(coast_rotated) + geom_sf()
}
}
