% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-index.R
\name{sim_cpue}
\alias{sim_cpue}
\title{Simulate fake commercial CPUE data for examples and testing}
\usage{
sim_cpue(sigma = 0.4, n_samples = 20, n_years = 15, n_vessels = 8)
}
\arguments{
\item{sigma}{TODO}

\item{n_samples}{TODO}

\item{n_years}{TODO}

\item{n_vessels}{TODO}
}
\description{
Simulate fake commercial CPUE data for examples and testing
}
\examples{
sim_cpue()
}
