% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-grid.R, R/surveys-spatial.R
\name{make_prediction_grid}
\alias{make_prediction_grid}
\alias{survey-spatial-modelling}
\alias{tidy_survey_sets}
\alias{interp_survey_bathymetry}
\alias{scale_survey_predictors}
\alias{fit_glmmfields}
\alias{fit_survey_sets}
\title{Survey spatial modelling}
\usage{
make_prediction_grid(dat, cell_width = 2, survey = NULL,
  cache_folder = "prediction-grids", utm_zone = 9, survey_boundary = NULL)

tidy_survey_sets(dat, survey, years, utm_zone = 9,
  density_column = "density_kgpm2")

interp_survey_bathymetry(dat, utm_zone = 9)

scale_survey_predictors(dat)

fit_glmmfields(dat, formula_positive = density ~ depth_scaled + depth_scaled2,
  formula_binary = present ~ depth_scaled + depth_scaled2, n_knots = 20,
  iter = 1000, chains = 4, adapt_delta = 0.98,
  cores = parallel::detectCores(), ...)

fit_survey_sets(dat, years, survey = NULL, density_column = "density_kgpm2",
  chains = 4, iter = 1000, max_knots = 20, adapt_delta = 0.95,
  thin = 1, mcmc_posterior_samples = 150, required_obs_percent = 0.05,
  utm_zone = 9, model = c("glmmfields", "inla"), include_depth = TRUE,
  survey_boundary = NULL, ...)
}
\arguments{
\item{dat}{TODO}

\item{cell_width}{Cell width in units of the data (e.g. UTM km).}

\item{survey}{TODO}

\item{cache_folder}{TODO}

\item{utm_zone}{TODO}

\item{years}{TODO}

\item{formula_positive}{TODO}

\item{formula_binary}{TODO}

\item{n_knots}{TODO}

\item{iter}{TODO}

\item{chains}{TODO}

\item{adapt_delta}{TODO}

\item{cores}{TODO}

\item{...}{TODO}

\item{max_knots}{TODO}

\item{thin}{TODO}

\item{mcmc_posterior_samples}{TODO}

\item{required_obs_percent}{TODO}

\item{region}{Optional region abbreviation to search for a spatial polygon
file. TODO}

\item{prediction_grid_n}{TODO}
}
\description{
Long description here
}
\details{
\itemize{
\item \code{tidy_survey_sets()} does...
\item \code{interp_survey_bathymetry()} does...
... TODO
}
}
\examples{
\dontrun{
## generally, use the main function:
x <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "Queen Charlotte Sound Synoptic Survey",
  iter = 600, chains = 1, mcmc_posterior_samples = 100)
names(x)
print(x$models)
plot_survey_sets(x$predictions, x$data, fill_column = "combined")

## internally, fit_survey_sets() does something like this:
dat <- tidy_survey_sets(pop_surv,
 survey = "SYN QCS",
 years = 2015)
dat_interp <- interp_survey_bathymetry(dat)
dat_scaled <- scale_survey_predictors(dat_interp$data)
m <- fit_glmmfields(dat_scaled)
pg <- make_prediction_grid(dat_scaled, survey = "SYN QCS")
pos <- predict(m$pos, newdata = data.frame(pg, time = 1),
 type = "response", return_mcmc = TRUE, iter = 100)
bin <- predict(m$bin, newdata = data.frame(pg, time = 1),
 type = "response", return_mcmc = TRUE, iter = 100)
com <- bin * pos
pg$combined <- apply(com, 1, median)
plot_survey_sets(pg, dat, fill_column = "combined")
}
}
\seealso{
Other spatial survey modelling functions: \code{\link{plot_survey_sets}}
}
\concept{spatial survey modelling functions}
