% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{fit_survey_sets}
\alias{fit_survey_sets}
\title{Spatial modeling of survey data}
\usage{
fit_survey_sets(
  dat,
  years,
  survey = NULL,
  density_column = "density_kgpm2",
  required_obs_percent = 0.05,
  utm_zone = 9,
  include_depth = TRUE,
  survey_boundary = NULL,
  premade_grid = NULL,
  tmb_knots = 200,
  cell_width = 2,
  ...
)
}
\arguments{
\item{dat}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}.}

\item{years}{The year to include in the model. Should be a single year.}

\item{survey}{The survey abbreviation. Should match the contents of the
column \code{survey_abbrev} in the data frame returned by \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}.}

\item{density_column}{The name of the column that includes the relative
biomass density to use. E.g. \code{"density_kgpm2"} for trawl surveys or
\code{"density_ppkm2"} for the long line surveys.}

\item{required_obs_percent}{A required fraction of positive sets before a
model is fit.}

\item{utm_zone}{The UTM zone to perform the modeling in. Defaults to zone 9.}

\item{include_depth}{Logical: should depth be included as a predictor? If
\code{FALSE} then the model will only have a spatial random field as the
predictor.}

\item{survey_boundary}{If not \code{NULL}, a data frame with the survey boundary
defined in columns \code{X} and \code{Y} in longitude and latitude coordinates. If
\code{NULL}, the functions will search for a matching element in the included
the data object \code{gfplot::survey_boundaries} based on the \code{survey} argument
(after removing "SYN" from the name).}

\item{premade_grid}{If not \code{NULL}, a list object with an element \code{grid} that
contains a data frame with columns \code{X}, \code{Y}, and \code{depth}, and another
element \code{cell_area} the content a single numeric value describing the grid
size in kilometers. The package includes a survey grid for the HBLL surveys
in \code{gfplot::hbll_grid}.}

\item{tmb_knots}{The number of knots to pass to \code{sdmTMB::sdmTMB()}.}

\item{cell_width}{The cell width if a prediction grid is made on the fly.}

\item{...}{Any other arguments to pass on to the modelling function.}
}
\description{
Implements geostatistical models of trawl or longline survey data.
}
\examples{
\dontrun{
set.seed(123)
# pop_surv <- gfdata::get_survey_sets("pacific ocean perch")
# or use built-in data:
fit <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "SYN QCS")
names(fit)
plot_survey_sets(fit$predictions, fit$data, fill_column = "combined")
}
}
