% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{fit_survey_sets}
\alias{fit_survey_sets}
\title{Spatial modeling of survey data}
\usage{
fit_survey_sets(
  dat,
  years,
  survey = NULL,
  density_column = "density_kgpm2",
  chains = 4,
  iter = 1000,
  max_knots = 20,
  adapt_delta = 0.95,
  thin = 1,
  mcmc_posterior_samples = 150,
  required_obs_percent = 0.05,
  utm_zone = 9,
  model = c("inla", "sdmTMB", "glmmfields"),
  include_depth = TRUE,
  survey_boundary = NULL,
  premade_grid = NULL,
  tmb_knots = 200,
  inla_knots_pos = 75,
  inla_knots_bin = 100,
  gamma_scaling = 1000,
  cell_width = 2,
  ...
)
}
\arguments{
\item{dat}{Output from \code{\link[=get_survey_sets]{get_survey_sets()}}.}

\item{years}{The year to include in the model. Should be a single year.}

\item{survey}{The survey abbreviation. Should match the contents of the
column \code{survey_abbrev} in the data frame returned by \code{\link[=get_survey_sets]{get_survey_sets()}}.}

\item{density_column}{The name of the column that includes the relative
biomass density to use. E.g. \code{"density_kgpm2"} for trawl surveys or
\code{"density_ppkm2"} for the long line surveys.}

\item{chains}{The number of MCMC chains. Only applies to the glmmfields
model.}

\item{iter}{The number of MCMC chains. Only applies to the glmmfields model.}

\item{max_knots}{The maximum number of knots to use in the predictive process
approximation in the glmmfields model. If this number is larger than the
number of data points then the number of knots is set to the number of data
points minus 2. Only applies to the glmmfields model.}

\item{adapt_delta}{Value to pass to \code{\link[rstan:sampling]{rstan::sampling()}}. Values closer to 1
make smaller steps in the Hamiltonian MCMC at the expense of speed. Only
applies to the glmmfields model.}

\item{thin}{Value to pass to \code{\link[rstan:sampling]{rstan::sampling()}} to thin the MCMC samples.
Only applies to the glmmfields model.}

\item{mcmc_posterior_samples}{Number of final MCMC samples to return. Applies
to both glmmfields and INLA.}

\item{required_obs_percent}{A required fraction of positive sets before a
model is fit.}

\item{utm_zone}{The UTM zone to perform the modeling in. Defaults to zone 9.}

\item{model}{The backend software to fit the model. Options are \code{"inla"} or
\code{"glmmfields"} (currently disabled until on CRAN). INLA implements in
approximation of the posterior via the integrated nested Laplace
approximation, but the results are usually quite good and are likely drive
considerably faster than glmmfields, especially for larger data sets or for
many knots.}

\item{include_depth}{Logical: should depth be included as a predictor? If
\code{FALSE} then the model will only have a spatial random field as the
predictor. Currently only applies to the INLA model.}

\item{survey_boundary}{If not \code{NULL}, a data frame with the survey boundary
defined in columns \code{X} and \code{Y} in longitude and latitude coordinates. If
\code{NULL}, the functions will search for a matching element in the included
the data object \code{gfplot::survey_boundaries} based on the \code{survey} argument
(after removing "SYN" from the name).}

\item{premade_grid}{If not \code{NULL}, a list object with an element \code{grid} that
contains a data frame with columns \code{X}, \code{Y}, and \code{depth}, and another
element \code{cell_area} the content a single numeric value describing the grid
size in kilometers. The package includes a survey grid for the HBLL surveys
in \code{gfplot::hbll_grid}.}

\item{tmb_knots}{The number of knots to pass to \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{inla_knots_pos}{The number of knots for the positive component model if
fit with INLA.}

\item{inla_knots_bin}{The number of knots for the binary component model if
fit with INLA.}

\item{gamma_scaling}{A value to multiply the positive densities by internally
before fitting the Gamma GLMM. The predictions are then divided by this
value internally to render a prediction on the original scale. The reason
for this is that values to are too small can create computational problems
for INLA.}

\item{cell_width}{The cell width if a prediction grid is made on the fly.}

\item{...}{Any other arguments to pass on to the modelling function.}
}
\description{
Implements geostatistical models of trawl or longline survey data. Uses INLA
or glmmfields (currently disabled) for Bayesian inference.
}
\examples{
\donttest{
set.seed(123)
# pop_surv <- get_survey_sets("pacific ocean perch")
# or use built-in data:
fit <- fit_survey_sets(pop_surv,
  years = 2015,
  survey = "SYN QCS")
names(fit)
plot_survey_sets(fit$predictions, fit$data, fill_column = "combined")
}
}
