% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageing-precision.R
\name{tidy_ageing_precision}
\alias{tidy_ageing_precision}
\title{Tidy PBS ageing precision data}
\usage{
tidy_ageing_precision(dat, ageing_method_codes = c(3, 17))
}
\arguments{
\item{dat}{A data frame from \code{\link{get_ageing_precision}}}

\item{ageing_method_codes}{A numeric vector of ageing method codes to filter
on. Defaults to codes \code{3} and \code{17}, which represent otolith
'break and burn' and 'break and bake' methods.}
}
\description{
Tidy PBS ageing precision data
}
\examples{
\dontrun{
get_ageing_precision("shortraker rockfish") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision()

get_ageing_precision("pacific ocean perch") \%>\%
  tidy_ageing_precision() \%>\%
  plot_ageing_precision(n = 200)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_ages}},
  \code{\link{tidy_bioindex}}, \code{\link{tidy_catch}},
  \code{\link{tidy_cpue_index}}

Other ageing precision functions: \code{\link{get_ageing_precision}},
  \code{\link{get_cpue_index}},
  \code{\link{plot_ageing_precision}}
}
