% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioindex.R
\name{tidy_bioindex}
\alias{tidy_bioindex}
\title{Title}
\usage{
tidy_bioindex(dat, surveys = c("West Coast Haida Gwaii Synoptic Survey",
  "Hecate Strait Synoptic Survey", "Queen Charlotte Sound Synoptic Survey",
  "West Coast Vancouver Island Synoptic Survey",
  "PHMA Rockfish Longline Survey - Outside North",
  "PHMA Rockfish Longline Survey - Outside South",
  "IRF Longline Survey (North)", "IRF Longline Survey (South)",
  "Hecate Strait Multispecies Assemblage Survey", "IPHC Longline Survey"),
  survey_names = c("West Coast Haida Gwaii Synoptic",
  "Hecate Strait Synoptic", "Queen Charlotte Sound Synoptic",
  "West Coast Vancouver Island Synoptic", "PHMA Rockfish LL - North",
  "PHMA Rockfish LL - South", "IRF LL Survey North", "IRF LL Survey South",
  "Hecate Strait Multispecies Assemblage", "IPHC Longline"), min_years = 3,
  year_range = NULL)
}
\arguments{
\item{dat}{TODO}

\item{surveys}{TODO}

\item{survey_names}{TODO}

\item{min_years}{TODO}

\item{year_range}{TODO}
}
\description{
Title
}
\examples{
\dontrun{
get_bioindex("lingcod") \%>\%
  tidy_bioindex() \%>\%
  plot_bioindex()

# Or without pipes:
d <- get_bioindex("lingcod")
head(d)

d_tidy <- tidy_bioindex(d)
head(d_tidy)

plot_bioindex(d_tidy)
}
}
\seealso{
Other tidy data functions: \code{\link{tidy_ageing_precision}},
  \code{\link{tidy_ages}}, \code{\link{tidy_catch}},
  \code{\link{tidy_cpue_index}}

Other survey biomass index functions: \code{\link{plot_bioindex}}
}
\concept{survey biomass index functions}
\concept{tidy data functions}
