% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{tidy_cpue_index}
\alias{tidy_cpue_index}
\title{Tidy a delta-lognormal commercial CPUE standardization model}
\usage{
tidy_cpue_index(object, center = TRUE)
}
\arguments{
\item{dat}{Input data frame}
}
\description{
Tidy a delta-lognormal commercial CPUE standardization model
}
