% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{tidy_cpue_index}
\alias{tidy_cpue_index}
\title{Tidy commercial PBS CPUE data}
\usage{
tidy_cpue_index(dat, species_common, year_range = c(1996, Inf),
  lat_range = c(48, Inf), min_positive_tows = 100, min_positive_trips = 4,
  min_yrs_with_trips = 4, area_grep_pattern = "5[CDE]+",
  lat_bands = seq(48, 59, 0.1), depth_bands = seq(50, 450, 25),
  gear = "BOTTOM TRAWL")
}
\arguments{
\item{dat}{An input data frame from \code{\link{get_cpue_index}}}

\item{species_common}{The species common name}

\item{year_range}{The range of years to include}

\item{lat_range}{The range of latitudes to include}

\item{min_positive_tows}{The minimum number of positive tows over all years}

\item{min_positive_trips}{The minimum number of annual positive trips}

\item{min_yrs_with_trips}{The number of years in which the
\code{min_positive_trips} criteria needs to be met}

\item{area_grep_pattern}{A regular expression to extract the management areas
of interest}

\item{lat_bands}{A sequence of latitude bans}

\item{depth_bands}{A sequence of depth bands}

\item{gear}{Gear types}
}
\description{
Tidy commercial PBS CPUE data
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
m <- fit_cpue_index(walleye)

plot_cpue_index_coefs(m)

predict_cpue_index(m) \%>\%
  plot_cpue_index()

jackknife_cpue_index(m)
}
}
\seealso{
Other tidy data functions: \code{\link{plot_surv_index}},
  \code{\link{tidy_age_precision}}

Other CPUE index functions: \code{\link{fit_cpue_index}},
  \code{\link{jackknife_cpue_index}},
  \code{\link{plot_cpue_index_coefs}},
  \code{\link{plot_cpue_index}},
  \code{\link{predict_cpue_index}}
}
\concept{CPUE index functions}
\concept{tidy data functions}
