% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-cpue-index.R
\name{tidy_cpue_index}
\alias{tidy_cpue_index}
\title{Tidy commercial PBS CPUE data}
\usage{
tidy_cpue_index(dat, species_common, year_range = c(1996,
  as.numeric(format(Sys.Date(), "\%Y")) - 1), lat_range = c(48, Inf),
  min_positive_tows = 100, min_positive_trips = 4, min_yrs_with_trips = 4,
  area_grep_pattern = "5[CDE]+", lat_band_width = 0.2,
  depth_band_width = 50, clean_bins = TRUE, depth_bin_quantiles = c(0.01,
  0.99), lat_bin_quantiles = c(0.01, 0.99), gear = "BOTTOM TRAWL")
}
\arguments{
\item{dat}{An input data frame from \code{\link[=get_cpue_index]{get_cpue_index()}}.}

\item{species_common}{The species common name.}

\item{year_range}{The range of years to include.}

\item{min_positive_tows}{The minimum number of positive tows over all years.}

\item{min_positive_trips}{The minimum number of annual positive trips.}

\item{min_yrs_with_trips}{The number of years in which the
\code{min_positive_trips} criteria needs to be met.}

\item{area_grep_pattern}{A regular expression to extract the management areas
of interest.}

\item{lat_band_width}{The latitude bandwidths in degrees.}

\item{depth_band_width}{The depth band widths in m.}

\item{clean_bins}{Logical. Should the depth and latitude bands be rounded to
the nearest clearn value as defined by \code{lat_band_width} or
\code{depth_band_width}? Internally, these use, for example:
\code{gfplot:::round_down_even(lat_range[1], lat_band_width)}.}

\item{depth_bin_quantiles}{Quantiles for the depth bands. Values above and
below these quantiles will be discarded.}

\item{lat_bin_quantiles}{Quantiles for the latitude bands. Values above and
below these quantiles will be discarded.}

\item{gear}{Gear types}
}
\description{
This function determines the qualifying "fleet" for CPUE analyses.
}
\examples{
\dontrun{
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
}
}
\seealso{
Other tidy data functions: \code{\link{plot_survey_index}},
  \code{\link{tidy_age_precision}}
}
\concept{tidy data functions}
