% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-props.R
\name{weight_comps}
\alias{weight_comps}
\title{Weight age or length composition data}
\usage{
weight_comps(dat)
}
\arguments{
\item{dat}{A properly formatted data frame. See Details. TODO}
}
\description{
Weight age or length composition data
}
