% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-props.R
\name{weight_comps}
\alias{weight_comps}
\alias{tidy_comps_commercial}
\alias{tidy_comps_survey}
\title{Age and length frequency weighting}
\usage{
tidy_comps_commercial(specimen_dat, catch_dat, value, bin_size = NULL)

tidy_comps_survey(specimen_dat, survey_tows, value, bin_size = NULL)

weight_comps(dat)
}
\arguments{
\item{specimen_dat}{Specimen data. E.g. from \code{\link[=get_comm_samples]{get_comm_samples()}} for
commercial data or \code{\link[=get_survey_samples]{get_survey_samples()}} for survey data.}

\item{catch_dat}{Catch data. E.g. from \code{\link[=get_catch]{get_catch()}}.}

\item{value}{The \strong{unquoted} column name with the values to re-weight
(e.g. \code{age} or \code{length}).}

\item{bin_size}{The binning size (likely only used for lengths).}

\item{survey_tows}{Survey tow data. E.g. from \code{\link[=get_survey_samples]{get_survey_samples()}}.}

\item{dat}{A properly formatted data frame such the output from
\code{tidy_comps_survey()} or \code{tidy_comps_commercial()}. See details.}
}
\description{
In general, you shouldn't have to use these functions directly. However, they
are exposed to the user in case you need the flexibility of using them
yourself and as a place to hold documentation about the weighting schemes.
These functions weight age or length frequencies. \code{tidy_comps_commercial()}
and \code{tidy_comps_survey()} join the necessary data for commercial or survey
samples, respectively, and format it for weighting. \code{weight_comps()} does the
actual weighting based on the output from the \code{tidy_*} functions.
}
\details{
The input data frame to \code{weight_comps()} must have columns in the
following order:
\enumerate{
\item year
\item weighting unit ID (e.g. trip ID or sample ID)
\item grouping variable for first weighting (e.g. quarter or stratum)
\item value of weighting variable (e.g. age or length bin)
\item frequency of that weighting variable (e.g. frequency of that age or
length bin)
\item numerator in first weighting (e.g. sample catch weight or density)
\item denominator in first weighting (e.g. quarter catch weight or total
stratum density)
\item numerator in second weighting (e.g. catch that quarter or stratum area)
\item denominator in second weighting (e.g. catch that year or total survey
area)
}

\code{tidy_comps_commercial()} and \code{tidy_comps_survey()} both output data frames
in this format and so you ordinarily would not have to worry about this. The
names of the columns do not matter, only their contents. The data frame
should contain these columns and only these columns.
}
\examples{
\dontrun{
species <- "redstripe rockfish"

## Surveys:
## ssid = 1 is Queen Charlotte Sound Synoptic Survey:
# rs_survey_samples <- get_survey_samples(species, ssid = 1)
# rs_survey_sets <- get_survey_sets(species, ssid = 1)

surv_lengths <- tidy_comps_survey(rs_survey_samples, rs_survey_sets,
  value = length, bin_size = 2)
surv_lengths
weight_comps(surv_lengths)

surv_ages <- tidy_comps_survey(rs_survey_samples, rs_survey_sets, value = age)
surv_ages
weight_comps(surv_ages)

## Commercial:
# rs_comm_samples <- get_comm_samples(species)
# rs_catch <- get_catch(species)

com_lengths <- tidy_comps_commercial(rs_comm_samples, rs_catch,
  value = length, bin_size = 2)
com_lengths
weight_comps(com_lengths)

com_ages <- tidy_comps_commercial(rs_comm_samples, rs_catch,
  value = age)
com_ages
weight_comps(com_ages)

## These functions are pipe (\%>\%) friendly. E.g.:
tidy_comps_survey(rs_survey_samples, rs_survey_sets, value = age) \%>\%
  weight_comps()
}
}
\references{
Page 161 of Kendra R. Holt, Paul J. Starr, Rowan Haigh, and Brian Krishka.
2016. Stock Assessment and Harvest Advice for Rock Sole (Lepidopsetta spp.)
in British Columbia. CSAS Res. Doc. 2016/009.
\href{http://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2016/2016_009-eng.html}{Link}
\href{http://waves-vagues.dfo-mpo.gc.ca/Library/363948.pdf}{PDF}
}
\seealso{
Other age- and length-frequency functions: \code{\link{plot_ages}},
  \code{\link{plot_lengths}}
}
\concept{age- and length-frequency functions}
